% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_process.R
\name{as.data.frame.dream_sequence}
\alias{as.data.frame.dream_sequence}
\title{Coerce a \code{dream_sequence} Object into a \code{data.frame} Object}
\usage{
\method{as.data.frame}{dream_sequence}(x, row.names = NULL, optional = FALSE, all.events = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{dream_sequence}.}

\item{row.names}{The \code{row.names} argument from the \code{\link[base]{as.data.frame}} function. Please
see that function for more details on this argument.}

\item{optional}{The \code{optional} argument from the \code{\link[base]{as.data.frame}} function. Please
see that function for more details on this argument.}

\item{all.events}{TRUE/FALSE. If sampling from the observed event sequence has occurred, TRUE
returns all of the sampled and non-sampled observed events, whereas FALSE returns only those
sampled observed events. FALSE by default.}

\item{...}{Additional arguments for other methods..}
}
\value{
Returns a \code{data.frame} object that contains the processed relational event statistics and
the associated sufficient network statistics for the event sequence.
}
\description{
This function will create a \code{data.frame} object from a \code{dream_sequence} object, where
the generated data.frame includes the processed event sequence and the
computed statistics. If \code{all.events} is set to FALSE, then the created
object will only contains the sampled events, whereas, if it is TRUE, the
returned object will contain all events, and if any computed statistics are attached,
the non-sampled event entries will be NAs.
}
\examples{

#a pseudo event sequence
events <- data.frame(time = 1:18, eventID = 1:18,
                     sender = c("A", "B", "C",
                                "A", "D", "E",
                                "F", "B", "A",
                                "F", "D", "B",
                               "G", "B", "D",
                               "H", "A", "D"),
                   target = c("B", "C", "D",
                               "E", "A", "F",
                               "D", "A", "C",
                                "G", "B", "C",
                                "H", "J", "A",
                                "F", "C", "B"))
#making a post-processing event sequence
processed <- create_res(type = "one-mode",
                       ordinal = TRUE,
                       riskset = "constant_sample",
                       time = events$time,
                       sender = events$sender,
                       receiver = events$target,
                       p_samplingobserved = 1.00,
                       n_controls = 1,
                       seed = 9999)

#computing the sender indegree statistics
processed <- dreamstats_degree(formation = "sender-indegree", data = processed)
 #computing the outgoing two paths statistics
processed <- dreamstats_triads(formation = "OTP", data = processed)
#computing the repetition/inertia statistics
processed <- dreamstats_repetition(data = processed)
#making the dream_sequence object a data.frame object
new.data <- as.data.frame(processed) #making the event set a data.frame object
new.data #the processed event sequence returned with the computed statistics
}
