% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_drug_expression_matrices.R
\name{build_drug_expression_matrices}
\alias{build_drug_expression_matrices}
\title{Converts expression matrices from L1000 data in tab delimited format.}
\usage{
build_drug_expression_matrices(
  L1000_data,
  L1000_metadata,
  include_controls = FALSE,
  L1000_data_CTRL = NULL,
  L1000_metadata_CTRL = NULL,
  dest_path = ".",
  inputgenes
)
}
\arguments{
\item{L1000_data}{An object of class gct or gctx from the LINCS1000 study (e.g. parsed by the parse.gct function of the cmapR package).}

\item{L1000_metadata}{Object containing the metadata relative to the L1000 dataset (e.g file GSE70138_Broad_LINCS_sig_info_2017-03-06.txt).}

\item{include_controls}{A logical value indicating whether to include the plate-specific control samples in the expression matrix.}

\item{L1000_data_CTRL}{An object of class gct or gctx from the LINCS1000 study for control samples (e.g. DMSO).}

\item{L1000_metadata_CTRL}{Object containing the metadata relative to the L1000 dataset for control samples.}

\item{dest_path}{Folder to be created as a destination folder for the expression matrices. Default to current directory.}

\item{inputgenes}{A character vector of gene symbols to be included in the expression matrix (e.g. L1000 landmark genes). Default all the genes will be considered.}
}
\value{
A squared matrix reporting the statistical significance of the similarity measures (e.g. Jaccard indices) among couples of drug targets neighborhoods.
}
\description{
This function converts expression matrices from the LINCS1000 dataset.
It provides flexibility in including control samples and allows selection of specific genes for the resulting expression matrix.
}
