% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlation_matrix.R
\name{calculate_correlation_matrix}
\alias{calculate_correlation_matrix}
\title{Infer correlation matrix from gene expression table using mutual information.}
\usage{
calculate_correlation_matrix(
  gx_table,
  iMethods,
  iEst,
  iDisc,
  ncores = 2,
  debug_output = FALSE
)
}
\arguments{
\item{gx_table}{A data frame representing the gene expression table.}

\item{iMethods}{A vector of valid inference algorithms for the MINET package.}

\item{iEst}{A vector of valid correlation methods for the MINET package.}

\item{iDisc}{A vector of valid discretization methods for the MINET package.}

\item{ncores}{The number of cores to use for running instances of MINET in parallel. Default: 2.}

\item{debug_output}{If TRUE, it prints help and status messages to aid in debugging the function. Default: FALSE.}
}
\value{
A binary symmetric matrix representing the median of mutual information correlations computed across various MINET combinations
}
\description{
This function, calculate_correlation_matrix, utilizes the MINET package to create a correlation matrix based on the mutual information method.
It allows the user to specify the inference algorithms, correlation calculation methods, and discretization methods, enabling multiple combinations of parameters for multiple runs of minet().
The resulting correlation matrices are combined using the median to create a consensus matrix.
}
