% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicals_distance.R
\name{chemicals_distance}
\alias{chemicals_distance}
\title{Calculate pairwise chemical distances between SMILES strings}
\usage{
chemicals_distance(
  smiles,
  method = "fingerprint",
  only_heavy_atoms = TRUE,
  radius = 4,
  n_bits = 2048,
  api_endpoint = "http://127.0.0.1:5000/chemicals_distance",
  python = "python"
)
}
\arguments{
\item{smiles}{A named list of SMILES strings.}

\item{method}{A character string specifying the method to use for calculating chemical distances.
"levenshtein", "fingerprint" and "mcs" are supported. Default is "fingerprint".}

\item{only_heavy_atoms}{A logical indicating whether to use only heavy atoms when calculating the MCS.
This parameter is only used when method = "mcs". Default is TRUE.}

\item{radius}{An integer specifying the radius for the ECFP fingerprint. This parameter is only used when
method = "fingerprint". Default is 4.}

\item{n_bits}{An integer specifying the length of the bit vector for the ECFP fingerprint. This parameter is
only used when method = "fingerprint". Default is 2048.}

\item{api_endpoint}{A character string specifying the URL of the API endpoint to use for calculating
chemical distances. Default is "http://127.0.0.1:5000/chemicals_distance".}

\item{python}{A character string specifying the path to the Python executable to use. Default is "python".}
}
\value{
A data.frame containing three columns: drug1, drug2, and distance. The drug1 and drug2 columns contain
the names of the two drugs being compared, and the distance column contains the calculated
chemical distance between the two drugs.
}
\description{
Given a list of named SMILES strings, this function calculates pairwise
chemical distances between all possible pairs
}
