% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_process.R
\name{create_res}
\alias{create_res}
\title{Process One- and Two-Mode Relational Event Sequences and Create Post-Processing Relational Event Sequences}
\usage{
create_res(
  type = c("two-mode", "one-mode"),
  ordinal = TRUE,
  t = NULL,
  time,
  sender,
  receiver,
  riskset = c("complete", "constant_sample", "dynamic_sample", "actor_varying",
    "actor_varying_sample"),
  p_samplingobserved = 1,
  n_controls = NULL,
  active_times = NULL,
  seed = NULL
)
}
\arguments{
\item{type}{"two-mode" indicates that this is a two-mode event sequence (i.e., observed actors
can only be either event senders or event receivers). The option "one-mode"
indicates that the observed event sequence is one-mode (i.e., observed actors
can be event senders and receivers)}

\item{ordinal}{TRUE/FALSE. TRUE indicates that observed timing of the events is ordinal (and the ordinal timing likelihood function will be used). FALSE denotes
that the observed timing is observed, relative to the start of the event sequence (and the interval timing likelihood function will be used). The interval
timing option adds the right-censored events to the post-processed relational event sequence (i.e., the set of (sampled) controls events for the
time point t, that marks the end of the relational event sequence.) Please see the references for more information.}

\item{t}{If ordinal is set to FALSE, the time that marks the end of the relational event sequence, relative to the start of the event
sequence. If t is NULL and ordinal is set to FALSE, then the right-censored events are not added to the post-processing event sequence and
the \code{\link{estimate_rem}} function will not add the right-censoring term to the log-likelihood. The log-likelihood thus
reduces to that used in the MLE interval timing relational event model in the \code{\link[remstimate]{remstimate}} function. The default
value is NULL.}

\item{time}{The vector of event time values from the observed event sequence,
where the jth entry is the relative time at which the jth event occurred. The event times
should be relative to the onset (start) of the relational event sequence.}

\item{sender}{The vector of event senders from the observed event sequence, where
the jth entry is the event sender for the jth observed/realized event.}

\item{receiver}{The vector of event receivers from the observed event sequence where
the jth entry is the event receiver for the jth observed/realized event.}

\item{riskset}{The argument should be one of the following strings: "complete", "constant_sample", "dynamic_sample",
"actor_varying", "actor_varying_sample". "complete" will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event is the full
set of actors that were active at anytime in the event sequence (for one-mode sequences, this is the
full Cartesian plot of actors, and for two-mode sequences, this is the full cross-product of the
event sender and receiver sets). "constant_sample" will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event is a random sample from the full
set of actors that were active at anytime in the event sequence (for one-mode sequences, this is the
full Cartesian plot of actors, and for two-mode sequences, this is the full cross-product of the
event sender and receiver sets), where the number of sampled events is dependent upon the \code{n_controls}
argument. "dynamic_sample" will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event at time t is a random sample from the full
set of actors that have been active up to and including t. "actor_varying" will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event at time t is the full
set of actors that are considered relationally active at time t dependent upon the \code{active_times}
argument, whereas the "actor_varying_sample" option will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event is a random
sample from the set of relationally active actors.}

\item{p_samplingobserved}{The numerical value for the probability of selection for sampling
from the observed event sequence. Set to 1 by default indicating that all observed
events from the event sequence will be included in the post-processing event sequence.}

\item{n_controls}{The numerical value for the number of null event controls for
each (sampled) observed event. This argument should be specified when one of the
following \code{riskset} types is provided: "constant_sample", "dynamic_sample",
and "actor_varying_sample".}

\item{active_times}{This argument is either a \code{data.frame} object for "one-mode" relational event
sequences or a \code{list} that contains 2 \code{data.frame} objects when \code{type} is set to "two-mode". For
"one-mode" event sequence types, the \code{data.frame} object must have three named
columns: "actor_id", "time_start", "time_end", and the number of rows (observations)
is the number of actor active spells, where an actor spell is defined as the
time in which the ith actor becomes relationally active (the ith entry for the "time_start"
column) and then becomes relationally inactive (the ith entry for the "time_end"
column). The "time_start" and "time_end" vectors should be relative
to the start of the relational event sequence (commonly set to start at time 0). Importantly, a single
actor (denoted by the "actor_id" column) can have  multiple active spell rows, as actors may exit and re-enter the relational event
sequence. In comparison, actors who do not leave the relational event sequence and always have the
capacity to become relationally active should have one row (entry) where the
"time_start" value is set to 0 and the "time_end" value should be set to the time point
that marks the end of the relational event sequence (t for interval event sequences
and the time of the last observed event for ordinal event sequences). Finally,
for "two-mode" event sequences, the \code{active_times} argument should be a list of
2 \code{data.frame} objects where the list elements are named "senders" and "receivers", where
"senders" is a \code{data.frame} object that follows the same design as the "one-mode"
arguments and defines how event senders become relationally active and inactive. Similarly,
the second element should be named "receivers" and is a \code{data.frame} object that follows the above
stated design and defines how event receivers become relationally active and inactive.}

\item{seed}{The random number seed for user replication. This argument is set to
NULL be default.}
}
\value{
An object of class \code{dream_sequence} that contains a list of the following elements:
\itemize{
\item \code{processed_sequence} - A \code{data.table} object that contains the post-processing relational event sequence with
the following columns: "time", "eventID", "sender", "receiver", "sampled", and "observed". The "time" column
is the vector of event times for the realized and control events. The "eventID" column represents the
order that the event occurred in the relational event sequence. The "sender" and "receiver" columns
are the specific dyad for that row. The "observed" vector takes a value of 1 if the
dyadic pair is the observed dyad at the specific time. The "sampled" vector takes a value
of 1 if the dyad was sampled at that event time and 0 if not (relevant for case-control sampling and sampling from the observed event sequence).
\item \code{ordinal} - Based upon the user's input. (\code{ordinal} and \code{interval})
\item \code{t} - Based upon the user's input.
\item \code{riskset} - Based upon the user's input.
\item \code{p} - The probability of sampling from the observed event sequence. Based upon the user's input.
\item \code{m} - The number of null event controls for each (sampled) observed event. Based upon the user's input.
\item \code{type} - Based upon the user's input. (\code{two-mode} and \code{one-mode})
\item \code{n} - The number of observed events.
\item \code{sampled_events} - The number of sampled observed events.
\item \code{null} - The number of sampled null (control) events.
\item \code{statistics} - An empty list to store the future computed relational event network statistics.
\item \code{interevent.times} - The vector of interevent times (the time difference between observed events).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function creates one- and two-mode post-processing (sampled) relational event
sequences with options for case-control sampling (Vu et al. 2015; Butts 2008), sampling
from the observed event sequence (Lerner and Lomi 2020), dynamic time-varying supports
sets, actor-varying supports sets, and the full Cartesian product for one-mode
sequences and the full cross-product for two-mode event sequences (Butts 2008). The created post-processing
relational event sequences are designed to be modeled by relational event models (Butts 2008).
Case-control sampling samples an arbitrary \emph{m} number of controls from the risk set for any event
(Vu et al. 2015; Butts 2008). Lerner and Lomi (2020) proposed sampling from the observed event sequence
where observed events are sampled with probability \emph{p}. Importantly, this function allows
users to generate post-processing relational event sequences for \emph{ordinal} and \emph{interval} relational
event model likelihoods. Lastly, the post-processing relational event sequence
is a \code{dream_sequence} object that is the required object for this package's
functions to compute exogenous and endogenous network statistics, alongside
the function to estimate Maximum Likelihood relational event models.
}
\details{
This function processes observed events from the set \eqn{A_t}, where each event \eqn{a_i} is
defined as:
\deqn{a_{i} \in A_t = (s_i, r_i, \tau_i, G[A_t; \tau_i])}
where:
\itemize{
\item \eqn{s_i} is the sender of the event.
\item \eqn{r_i} is the receiver of the event.
\item \eqn{\tau_i} represents the time of the event.
\item \eqn{G[A_t; \tau_i] = \{a_1, a_2, \ldots, a_{t'} \mid t' < \tau_i\}} is the network of past events, that is, all events that occurred prior to the current event, \eqn{a_i}.
}

For the post-processing event sequences where the \code{ordinal} argument is set to FALSE, the
last set of processed events, marked with the time point \eqn{t}, represent the
right-censoring events. The function generates post-processing relational event
sequences across three axises: (1) the inclusion of sampling from the observed/realized
relational event sequence, \eqn{A_t}, (2) one-mode vs. two-mode event types,
where the relevant actors can be either senders or receivers (in the case of two-mode)
sequences, or where the relevant actors can be both (in the case of one-mode
sequences), and (3) how the processed support set for each event should be
constructed. The third axis is based upon the \code{riskset} argument, which is
one of the following: "complete", "constant_sample", "dynamic_sample",
"actor_varying", "actor_varying_sample". "complete" will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event is the full
set of actors that were active at anytime in the event sequence (for one-mode sequences, this is the
full Cartesian plot of actors, and for two-mode sequences, this is the full cross-product of the
event sender and receiver sets). "constant_sample" will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event is a random sample from the full
set of actors that were active at anytime in the event sequence (for one-mode sequences, this is the
full Cartesian plot of actors, and for two-mode sequences, this is the full cross-product of the
event sender and receiver sets), where the number of sampled events is dependent upon the \code{n_controls}
argument. "dynamic_sample" will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event at time t is a random sample from the full
set of actors that have been active up to and including t. "actor_varying" will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event at time t is the full
set of actors that are considered relationally active at time t dependent upon the \code{active_times}
argument, whereas the "actor_varying_sample" option will create a post-processing relational event
sequence where the set of null events for each (sampled) realized/observed event is a random
sample from the set of relationally active actors.

\strong{Complete Risk Sets}:

Following Butts (2008) and Butts and Marcum (2017), for one-mode
event sequences (\code{type} = "one-mode"), the risk (support) set is defined as all possible
events at time \eqn{t}, \eqn{M_t}, as the full Cartesian
product of actors active in the relational event sequence and is the defined as the
set \eqn{N}.  Formally:
\deqn{M_t = \{ (s, r) \mid s \in N \times r \in N\}}
where \eqn{N} is the set of all possible actors in the sequence. In this function,
the full risk set is considered fixed (constant) across all time points.

For two-mode event sequences (\code{type} = "two-mode"), the risk (support) set is defined as all possible
events at time \eqn{t}, \eqn{M_t}, as the cross product of two disjoint sets, namely, event senders (i.e., \eqn{S}) and
event receivers (i.e., \eqn{R}). Formally:
\deqn{M_t = \{ (s, r) \mid s \in S \times r \in R\}}
where \eqn{S} is the set of potential event senders and \eqn{R} is the set of potential event receivers. In this function,
the full risk set is considered fixed across all time points.

\strong{Constant Sample Risk Sets}:

Following Butts (2008), Vu et al. (2015), and Lerner and Lomi (2020), case-control sampling
samples an arbitrary number \eqn{m} of non-events from the above risk/support set definitions \eqn{M_t}. This
process generates a new support set, \eqn{\tilde{M}_t}, for any relational event
\eqn{a_i} contained in \eqn{A_t}. \eqn{\tilde{M}_t}, for one-mode
relational event sequences, is formally defined as:
\deqn{\tilde{M}_t \subseteq \{ (s, r) \mid s \in N \times r \in N \}}

\strong{Dynamic Sample Risk Sets}:

For dynamic risk sets, for one-mode event sequences (\code{type} = "one-mode"), the risk (support)
set at time \eqn{t}, that is, \eqn{M_t}, is defined as a sample of \eqn{m} dyads from the full Cartesian product
of all past actors who have been involved in a relational event at or before time \eqn{t}.
Formally:
\deqn{\tilde{M}_t \subseteq \{ (s, r) \mid s \in N_t \times r \in N_t\}}
where \eqn{N_t} is the set of potential event senders and targets at and time \eqn{t}. The
definition follows the same as above for two-mode event sequences, where the sets
are now defined as \eqn{S_t} and \eqn{R_t}.

\strong{Actor-varying Risk Sets}:

Actor-varying support sets allows for actors to enter, exit, and re-enter
the relational event sequence as time progresses. For one-mode event sequences,
the node set \eqn{Y_t} is defined as those actors who are considered
relationally active at time \eqn{t}. For two-mode event sequences,
the node sets \eqn{S_t} and \eqn{R_t} are defined, respectively, as the
senders who can be active at time \eqn{t} and the receivers who can be
active at time \eqn{t}. For one-mode sequences, the formal definition is:
\deqn{V_t = \{ (s, r) \mid s \in Y_t \times r \in Y_t\}}

\strong{Actor-varying Sampling Risk Sets}:

Based upon the formal definition above, sampling from the full actor-varying
risk set generates a new support set definition:
\deqn{\tilde{V_t} \subseteq \{ (s, r) \mid s \in Y_t \times r \in Y_t\}}
where \eqn{\tilde{V_t}} represents the \eqn{m} number of dyads sampled
(with equal probability) from the set \eqn{V_t}.
}
\examples{

# Generating a psuedo one-mode relational event sequence
set.seed(9999)
events <- data.frame(time = sort(rexp(18)),
                                eventID = 1:18,
                                sender = c("A", "B", "C",
                                           "A", "D", "E",
                                           "F", "B", "A",
                                           "F", "D", "B",
                                           "G", "B", "D",
                                          "H", "A", "D"),
                               target = c("B", "C", "D",
                                          "E", "A", "F",
                                          "D", "A", "C",
                                          "G", "B", "C",
                                          "H", "J", "A",
                                          "F", "C", "B"))

# Creating the full one-mode relational risk set with p = 1.00 (all true events)
# based upon the ordinal timing relational event framework
full.process <- create_res(ordinal = TRUE,
                      type = "one-mode",
                      riskset = "complete",
                      time = events$time,
                      sender = events$sender,
                      receiver = events$target,
                      p_samplingobserved = 1.00,
                      seed = 9999)

# Creating a fixed one-mode relational risk set with p = 1.00 (all true events)
# and 5 controls based upon the ordinal timing relational event framework
sample.process <- create_res(ordinal = TRUE,
                            type = "one-mode",
                            riskset = "constant_sample",
                            time = events$time,
                            sender = events$sender,
                            receiver = events$target,
                            p_samplingobserved = 1.00,
                            n_controls = 10,
                            seed = 9999)

# Creating a dynamic one-mode relational risk set with p = 1.00 (all true events)
# and 5 controls based upon the interval timing relational event framework
dynamic.process <- create_res(ordinal = FALSE,
                      t = max(events$time) + rexp(1),
                      type = "one-mode",
                      riskset = "dynamic_sample",
                      time = events$time,
                      sender = events$sender,
                      receiver = events$target,
                      p_samplingobserved = 1.00,
                      n_controls = 5,
                      seed = 9999)


# Creating a actor-varying one-mode relational event sequence where actors
# enter, exit, and re-enter the event sequence dependent upon the user-specified
# active times. Each row contains the actor id, the time for which, in a specific relevant
#spell, become active (enter the sequence) and become inactive (exit the sequence)
#for actors who are always relevant, the active time ranges from 0 to the
#end of the sequence
t <- max(events$time) + rexp(1)
active.times <- data.frame(actor_id = c("A", "B", "B", "C", "D", "E", "F", "G",
                                       "H" ,"J"),
                         time_start = c(0,0,0.65,0,0,0.45,0.60,0,0,0),
                         time_end = c(t,0.45,t,t,t,1.00,t,t,t,t))

actor.varying.process <- create_res(ordinal = TRUE,
                                   type= "one-mode",
                                   riskset = "actor_varying",
                                   time = events$time,
                                   sender = events$sender,
                                   receiver = events$target,
                                   active_times = active.times)


}
\references{
Butts, Carter T. 2008. "A Relational Event Framework for Social Action." \emph{Sociological Methodology} 38(1): 155-200.

Butts, Carter T. and Christopher Steven Marcum. 2017. "A Relational Event Approach to Modeling Behavioral Dynamics." In A.
Pilny & M. S. Poole (Eds.), \emph{Group processes: Data-driven computational approaches}. Springer International Publishing.

Lerner, Jürgen and Alessandro Lomi. 2020. "Reliability of relational event model estimates under sampling: How to
fit a relational event model to 360 million dyadic events." \emph{Network Science} 8(1): 97–135.

Vu, Duy, Philippa Pattison, and Garry Robins. 2015. "Relational event models for social learning in MOOCs." \emph{Social Networks} 43: 121-135.
}
\author{
Kevin A. Carson \href{mailto:kacarson@arizona.edu}{kacarson@arizona.edu} and Diego F. Leal \href{mailto:dflc@arizona.edu}{dflc@arizona.edu}
}
