% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_dataframe.R
\name{distance_to_dataframe}
\alias{distance_to_dataframe}
\title{Convert a distance matrix to a data frame.}
\usage{
distance_to_dataframe(distance_matrix)
}
\arguments{
\item{distance_matrix}{A symmetric distance matrix.}
}
\value{
A data frame with three columns: "drug1", "drug2", and "distance", representing pairs of drugs and their corresponding distances.
}
\description{
This is an internal function that converts a distance matrix to a data frame format. It considers the distance matrix to be symmetric and only uses the upper triangle of the matrix.
}
