% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream-package.R
\docType{package}
\name{dream-package}
\alias{dream}
\alias{dream-package}
\title{dream: Dynamic Relational Event Analysis and Modeling}
\description{
A set of tools for relational and event analysis, including two- and one-mode network brokerage and structural measures, and helper functions optimized for relational event analysis with large datasets, including creating relational risk sets, computing network statistics, estimating relational event models, and simulating relational event sequences. For more information on relational event models, see Butts (2008) \doi{10.1111/j.1467-9531.2008.00203.x}, Lerner and Lomi (2020) \doi{10.1017/nws.2019.57}, Bianchi et al. (2024) \doi{10.1146/annurev-statistics-040722-060248}, and Butts et al. (2023) \doi{10.1017/nws.2023.9}. In terms of the structural measures in this package, see Leal (2025) \doi{10.1177/00491241251322517}, Burchard and Cornwell (2018) \doi{10.1016/j.socnet.2018.04.001}, and Fujimoto et al. (2018) \doi{10.1017/nws.2018.11}. This package was developed with support from the National Science Foundation’s (NSF) Human Networks and Data Science Program (HNDS) under award number 2241536 (PI: Diego F. Leal). Any opinions, findings, and conclusions, or recommendations expressed in this material are those of the authors and do not necessarily reflect the views of the NSF.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinCarson/dream}
  \item Report bugs at \url{https://github.com/kevinCarson/dream/issues}
}

}
\author{
\strong{Maintainer}: Kevin A. Carson \email{kacarson@arizona.edu} (\href{https://orcid.org/0009-0001-5762-2586}{ORCID})

Authors:
\itemize{
  \item Diego F. Leal \email{dflc@arizona.edu} (\href{https://orcid.org/0000-0001-9973-4626}{ORCID})
}

}
\keyword{internal}
