% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_process.R
\name{dream_sequence}
\alias{dream_sequence}
\title{Helper Function to Create \code{dream_sequence} Objects}
\usage{
dream_sequence(
  ordinal = TRUE,
  t = NULL,
  time,
  sender,
  receiver,
  observed = NULL,
  sampled = NULL,
  type = "one-mode",
  statistics = NULL,
  ...
)
}
\arguments{
\item{ordinal}{TRUE/FALSE. TRUE indicates that observed timing of the events is ordinal (and the ordinal timing likelihood function will be used). FALSE denotes
that the observed timing is observed, relative to the start of the event sequence (and the interval timing likelihood function will be used). Please see the references for more information.}

\item{t}{If ordinal is set to FALSE, the time that marks the end of the relational event sequence, relative
to the start of the event sequence.}

\item{time}{A numeric vector that contains the timing of the events in the relational event sequence.}

\item{sender}{A character vector that contains the sender of the events in the relational event sequence.}

\item{receiver}{A character vector that contains the receiver/target of the events in the relational event sequence.}

\item{observed}{A numeric vector that is 1 if the observation is an observed event in the relational event sequence, or 0 if the observation is a control event in the relational event sequence (see \code{\link{create_res}} for more information).}

\item{sampled}{A numeric vector that is 1 if the observation is a sampled event in the relational event sequence, or 0 if the observation is a non-sampled event in the relational event sequence (see \code{\link{create_res}} for more information).}

\item{type}{"two-mode" indicates that this is a two-mode event sequence. "one-mode" indicates that the event sequence is one-mode.}

\item{statistics}{A \code{data.frame} object that contains the previously computed statistics
for the processed events. If processing has occurred (i.e., the \code{sampled} argument is
specified; control events have been created and/or if sampling from the observed event sequence),
the the \code{statistics} argument must have a number of observations equal to the sum of the
\code{sampled} argument. If processing has not occurred, then the \code{statistics} argument must have
a number of observations equal to the length of the \code{time} argument.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{dream_sequence} object that contains the user-provided information.
}
\description{
The function \code{dream_sequence()} is a user helper function that transforms
user-created processed event sequences into \code{dream_sequence} objects to be
used in the \code{dream} functions to compute sufficient network statistics
and estimate relational event models. This function may also be helpful
to user who need to computed network statistics for the estimation of
relational outcome models (see \href{https://www.cambridge.org/core/journals/network-science/article/microlevel-network-dynamics-of-scientific-collaboration-and-impact-relational-hyperevent-models-for-the-analysis-of-coauthor-networks/375932B5B86D2033A0A290DE8198BB32}{Lerner and Hâncean (2023)}).

If \code{ordinal} is FALSE, that is, if the relational event sequence is to use the
interval timing likelihood, then the events for the last observation time point (the
set of realized and null events at the last time point) should all be control events, as
they represent the set of right-censoring non-realized events. The \code{t} should
specify the time point that marks the end of the relational event sequence \eqn{A_t}. If \code{t}
is not known, then the value should be left as \code{NULL}.
}
\examples{
#a pseudo event sequence
events <- data.frame(time = 1:18, eventID = 1:18,
                     sender = c("A", "B", "C",
                                "A", "D", "E",
                                "F", "B", "A",
                                "F", "D", "B",
                               "G", "B", "D",
                               "H", "A", "D"),
                   target = c("B", "C", "D",
                               "E", "A", "F",
                               "D", "A", "C",
                                "G", "B", "C",
                                "H", "J", "A",
                                "F", "C", "B"))
#making a post-processing event sequence
eventSet <- create_res(type = "one-mode",
                       ordinal = TRUE,
                       riskset = "fixed",
                       case_control=TRUE,
                       time = events$time,
                       sender = events$sender,
                       receiver = events$target,
                       p_samplingobserved = 1.00,
                       n_controls = 1,
                       seed = 9999)

#computing the sender indegree statistics
eventSet <- dreamstats_degree(formation = "sender-indegree", data = eventSet)
#making the dream_sequence object a data.frame object
new.data <- as.data.frame(eventSet) #making the event set a data.frame object
stats <- new.data["sender.indegree"] #the computed statistics

#reconverting the object to a dream_sequence object
psuedo.data <- dream_sequence(ordinal = TRUE,
                              time = new.data$time,
                              sender = new.data$sender,
                              receiver = new.data$receiver,
                              observed = new.data$observed,
                              sampled = new.data$sampled,
                              type= "one-mode",
                              statistics = stats)
psuedo.data #printing the object

#reconverting we original event sequence to a dream_sequence object
#(this is helpful for the estimation of relational outcome models!)
psuedo.data1 <-  dream_sequence(ordinal = TRUE,
                                time = events$time,
                                sender = events$sender,
                                receiver = events$target,
                                type= "one-mode")

psuedo.data1 #printing the object

#computing a statistic on the data
psuedo.data1 <- dreamstats_degree(formation = "sender-outdegree",
                                  data = psuedo.data1,
                                  counts = TRUE)

psuedo.data1 #printing the object with the computed statistics
psuedo.data1$statistics #printing the object with the computed statistics
}
