% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_remstats.R
\name{dreamstats_actor}
\alias{dreamstats_actor}
\title{Add Actor-Level Statistics for Event Dyads in a Relational Event Sequence}
\usage{
dreamstats_actor(
  data,
  var_name,
  sender_ref = TRUE,
  actor_info,
  make_factor = FALSE,
  return_stats = FALSE
)
}
\arguments{
\item{data}{A \code{dream_sequence} object that contains the processed relational event sequence.}

\item{var_name}{A string that is the name of the variable from the \code{actor_info} argument that represents the actor-level statistic to be added.}

\item{sender_ref}{TRUE/FALSE. TRUE indicates that the variable should be added with respect to the sender (i.e., the variable is associated with the sender). FALSE indicates that the variable is representative of the event receivers.}

\item{actor_info}{A \eqn{N} x 4 \code{data.frame} object that contains four columns with the number of observations
being the number of referenced actors (i.e., if sender = TRUE, then the senders in the relational event sequence). The object should contain
four named columns: (1) \code{actor_id}, which is the id that is associated with the actors, (2) \code{var_name} (from the argument), where
the var_name column is the vector of statistics, and for time-varying statistics, (3) \code{time_start}, a column that contains the
time that represents the time at which the actor has that statistic, and (4) \code{time_end}, a column that contains the
time for which the actor stops possessing that statistic value (see details for more information).}

\item{make_factor}{TRUE/FALSE. TRUE indicates that the vector of values will be made a factor, and FALSE if not.}

\item{return_stats}{TRUE/FALSE. TRUE indicates that the vector of computed
statistics will be returned. FALSE indicates that the vector of computed
statistics will be added to the \code{statistics} element of the \code{data} argument.}
}
\value{
The vector of actor-level statistics for the relational event sequence or the updated \code{data} argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function allows users to add time-varying and time-invariant
actor-level statistics to be used in the estimation of ordinal and interval timing relational event models.
}
\details{
This function adds user-provided time-varying and time-invariant actor-level
statistics to a \code{dream_sequence} object for relational event models. The \code{actor_info}
argument should be a \eqn{N} x 4 \code{data.frame} object with four named columns. The first
column should be named \code{actor_id}, which represents the sender/receiver unique ids
based upon the observed relational event sequence. The second column should be named
\code{time_start} and represents the time in which the specific actor adopts the specific
value. The third column should be named \code{time_end} and represents the time in which the specific actor adopts the specific
value. For time-invariant actor-level statistics, the \code{time_start} value should be
0 and the \code{time_end} value should be the time that marks the end of the relational
event sequence (e.g., the time of the last observed event). The last column should be named after the
\code{var_name} argument and represents the actor-level statistic for the actor i during the
times between \code{time_start} and \code{time_end}.
}
\examples{
events <- data.frame(time = 1:18, eventID = 1:18,
                    sender = c("A", "B", "C",
                               "A", "D", "E",
                               "F", "B", "A",
                               "F", "D", "B",
                               "G", "B", "D",
                               "H", "A", "D"),
                    target = c("B", "C", "D",
                               "E", "A", "F",
                              "D", "A", "C",
                               "G", "B", "C",
                               "H", "J", "A",
                               "F", "C", "B"))

processed <- create_res(type = "one-mode",
                      ordinal = TRUE,
                      riskset = "constant_sample",
                      time = events$time,
                      sender = events$sender,
                      receiver = events$target,
                      p_samplingobserved = 1.00,
                      n_controls = 1,
                      seed = 9999)

#reconstructing the data.frame object to store the time-varying
#actor-level statistic for sender
ids <- unique(c(events$sender,events$target))
actor_stats <- data.frame(actor_id = ids,
                         time_start = 0,
                         time_end = 18,
                         rv = sample(0:1, length(ids), TRUE))

#adding the value to the post-processing relational event sequence where
#the new variable is named "rv"
processed <- dreamstats_actor(data = processed,
                             var_name = "rv",
                             sender_ref = TRUE,
                             actor_info = actor_stats)
processed
processed$statistics$rv.sender

#constructing the data.frame object to store the time-invariant
#actor-level statistic for receivers
ids <- unique(c(events$sender,events$target))
actor_stats <- data.frame(actor_id = rep(ids,2),
                         time_start = c(rep(0, 9), rep(10,9)),
                         time_end = c(rep(9,9),rep(18,9)),
                         trv = rnorm(length(ids)*2))

#adding the value to the post-processing relational event sequence where
#the new variable is named "trv" and the values is for the event receivers
processed <- dreamstats_actor(data = processed,
                             var_name = "trv",
                             sender_ref = FALSE,
                             actor_info = actor_stats)
processed
processed$statistics$trv.receiver

extract.data <- as.data.frame(processed)
extract.data

}
\author{
Kevin A. Carson \href{mailto:kacarson@arizona.edu}{kacarson@arizona.edu}, Diego F. Leal \href{mailto:dflc@arizona.edu}{dflc@arizona.edu}
}
