% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_remstats.R
\name{dreamstats_actorfe}
\alias{dreamstats_actorfe}
\title{Add Actor-Level Fixed Effects for Event Dyads in a Relational Event Sequence}
\usage{
dreamstats_actorfe(data, sender = TRUE, return_stats = FALSE)
}
\arguments{
\item{data}{A \code{dream_sequence} object that contains the processed relational event sequence.}

\item{sender}{TRUE/FALSE. TRUE creates event sender fixed effects and FALSE created event receiver fixed effects.}

\item{return_stats}{TRUE/FALSE. TRUE indicates that the vector of computed
statistics will be returned. FALSE indicates that the vector of computed
statistics will be added to the \code{statistics} element of the \code{data} argument.}
}
\value{
The vector of actor-level fixed effects for the relational event sequence or the updated \code{data} argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function allows users to add event sender and receiver fixed effects
for relational event models to a \code{dream_sequence} object.
}
\details{
This function adds sender or receiver actor-level fixed effects
to a \code{dream_sequence} object. Internally, the function creates a new variable
(\code{senderFE} for sender fixed effects and \code{receiverFE} for receiver fixed effects)
that is a factor of the event sender/receiver ids.
}
\examples{
events <- data.frame(time = 1:18, eventID = 1:18,
                    sender = c("A", "B", "C",
                               "A", "D", "E",
                               "F", "B", "A",
                               "F", "D", "B",
                               "G", "B", "D",
                               "H", "A", "D"),
                    target = c("B", "C", "D",
                               "E", "A", "F",
                              "D", "A", "C",
                               "G", "B", "C",
                               "H", "J", "A",
                               "F", "C", "B"))

processed <- create_res(type = "one-mode",
                      ordinal = TRUE,
                      riskset = "constant_sample",
                      time = events$time,
                      sender = events$sender,
                      receiver = events$target,
                      p_samplingobserved = 1.00,
                      n_controls = 20,
                      seed = 9999)

#adding the sender fixed effects to the statistics list
processed <- dreamstats_actorfe(data=processed,sender=TRUE)

#adding the receiver fixed effects to the statistics list
processed <- dreamstats_actorfe(data=processed,sender=FALSE)

processed #the effects are added

#estimating the fixed effects only ordinal timing relational event model
model <- estimate_rem(~senderFE + receiverFE, data = processed)

}
\author{
Kevin A. Carson \href{mailto:kacarson@arizona.edu}{kacarson@arizona.edu}, Diego F. Leal \href{mailto:dflc@arizona.edu}{dflc@arizona.edu}
}
