% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_to_num.R
\name{fct_to_num}
\alias{fct_to_num}
\title{Factor to numeric}
\usage{
fct_to_num(x)
}
\arguments{
\item{x}{A factor.}
}
\value{
A numeric.
}
\description{
Easily convert a factor to numeric without losing information.
}
\examples{

library(dplyr)
set.seed(1)
num <- rnorm(10)


# Without fct_to_numeric
data.frame(fac = factor(num)) \%>\%
 mutate(bad_num = as.numeric(fac))

# With fct_to_numeric
data.frame(fac = factor(num)) \%>\%
 mutate(good_num = fct_to_num(fac))
}
