% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm.R
\name{genetic_algorithm}
\alias{genetic_algorithm}
\title{Apply a genetic algorithm to optimize drug combinations.}
\usage{
genetic_algorithm(
  smiles,
  moa,
  graph,
  disease_network,
  rank,
  targets,
  population_size = 5,
  n_offsprings = 20,
  attribute_init_prob = 0.3,
  attribute_mutation_prob = 0.1,
  crossover_prob = 0.7,
  mutation_prob = 0.3,
  n_generations = 10,
  api_endpoint = "http://127.0.0.1:5000/genetic_algorithm",
  python = "python"
)
}
\arguments{
\item{smiles}{a data frame with columns for drug 1, drug 2, and the distance between them based on their structure.}

\item{moa}{a data frame with columns for drug 1, drug 2, and the distance between them based on their mechanism of action (MOA)}

\item{graph}{a data frame with columns for drug 1, drug 2, and the distance between them based on the length of their shortest path on the disease-relevant gene network.}

\item{disease_network}{an interaction network of the disease-relevant genes.}

\item{rank}{a ranking of genes based on their degree in the disease-relevant gene network.}

\item{targets}{a data.frame in the format of OpenTargets specifying the target genes of each drug.}

\item{population_size}{an integer specifying the size of the population. Default is 5.}

\item{n_offsprings}{an integer specifying the number of offsprings produced in each generation. Default is 20.}

\item{attribute_init_prob}{a numeric value specifying the probability of choosing a drug during attribute initialization. Default is 0.3.}

\item{attribute_mutation_prob}{a numeric value specifying the probability of attribute mutation. Default is 0.1.}

\item{crossover_prob}{a numeric value specifying the probability of crossover. Default is 0.7.}

\item{mutation_prob}{a numeric value specifying the probability of mutation. Default is 0.3.}

\item{n_generations}{an integer specifying the number of generations to run. Default is 10.}

\item{api_endpoint}{the URL endpoint of the API. Default is "http://127.0.0.1:5000/genetic_algorithm".}

\item{python}{the path to the Python executable. Default is "python".}
}
\value{
a named list containing the following fields:
drug_names: a list with the names of the drugs considered
hall_of_fame: the list of best drug combinations identified during the execution of the algorithm, with their scores from the fitness function
logbook: the average score of each fitness function for each generation
population: a list of binary arrays representing the individuals in the population in the last generation of the algorithm.
}
\description{
This function applies a genetic algorithm to find the best combinations of
drugs to treat a disease, based on multiple distance metrics.
}
