% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_jacc_indices.R
\name{get_avg_jacc_indices}
\alias{get_avg_jacc_indices}
\title{Creates a squared matrix of average jaccard indices between subsets of neighbors of a certain order of couples of drugs. Note that each drug may have multiple targets, so the average shortest paths of all the targets of a certain drug is computed.}
\usage{
get_avg_jacc_indices(Graph, drug_target_df, drugs_col, targets_col, order = 1)
}
\arguments{
\item{Graph}{An object of class "igraph".}

\item{drug_target_df}{Dataframe of at least two columns containing drug names and corresponding targets.}

\item{drugs_col}{Column index of the dataframe drug_target_df where the drugs are stored.}

\item{targets_col}{Column index of the dataframe drug_target_df where the drugs are stored.}

\item{order}{Order of the neighborhood to be considered.}
}
\value{
A squared matrix of average shortest paths between couples of drug targets.
}
\description{
Creates a squared matrix of average jaccard indices between subsets of neighbors of a certain order of couples of drugs. Note that each drug may have multiple targets, so the average shortest paths of all the targets of a certain drug is computed.
}
