% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drug_avg_rank_from_graph.R
\name{get_drug_avg_rank_from_graph}
\alias{get_drug_avg_rank_from_graph}
\title{Creates a squared matrix of average jaccard indices between subsets of neighbors of a certain order of couples of drugs. Note that each drug may have multiple targets, so the average shortest paths of all the targets of a certain drug is computed.}
\usage{
get_drug_avg_rank_from_graph(
  drug_target_df,
  drugs_col,
  targets_col,
  graph_rank
)
}
\arguments{
\item{drug_target_df}{Dataframe of at least two columns containing drug names and corresponding targets.}

\item{drugs_col}{Column index of the dataframe drug_target_df where the drugs are stored.}

\item{targets_col}{Column index of the dataframe drug_target_df where the drugs are stored.}

\item{graph_rank}{An ordered named vector of a graph's nodes. Typically obtained from the get_graph_named_ranklist function.}
}
\value{
A rank of drugs based on the average ranks of their targets.
}
\description{
Creates a squared matrix of average jaccard indices between subsets of neighbors of a certain order of couples of drugs. Note that each drug may have multiple targets, so the average shortest paths of all the targets of a certain drug is computed.
}
