% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_named_ranklist.R
\name{get_graph_named_ranklist}
\alias{get_graph_named_ranklist}
\title{Get a named list from an igraph object using ranked vectors}
\usage{
get_graph_named_ranklist(Graph, ranked_vectors)
}
\arguments{
\item{Graph}{An igraph object representing the graph.}

\item{ranked_vectors}{A list of one or more ranked vectors.}
}
\value{
A named list based on the provided ranked vectors.
}
\description{
This function takes an igraph object and a list of ranked vectors as input, and generates a named list based on the provided ranked vectors.
The function assigns the ranked vectors as vertex attributes in the graph and then calculates a ranked gene list using specified network attributes.
The resulting ranked gene list is then returned as a named list
}
