% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iGraph.R
\name{get_iGraph}
\alias{get_iGraph}
\title{Create iGraph object from a symmetrical adjacency matrix, annotate it with centrality attributes and return the annotated iGraph.}
\usage{
get_iGraph(adj_mat)
}
\arguments{
\item{adj_mat}{Binary consensus matrix computed by using get_ranked_consensus_binary_matrix() function.}
}
\value{
Annotated igraph object representing the binary consensus matrix computed by get_ranked_consensus_binary_matrix() function.
}
\description{
Create iGraph object from a symmetrical adjacency matrix, annotate it with centrality attributes and return the annotated iGraph.
}
