% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_informative_genes.R
\name{get_informative_genes}
\alias{get_informative_genes}
\title{Identify informative genes for co-expression network construction.}
\usage{
get_informative_genes(
  expr_mat,
  group,
  test = "wilcoxon + var.test",
  percentile = 10
)
}
\arguments{
\item{expr_mat}{An expression matrix where rows are genes and columns are samples.}

\item{group}{A factor indicating the biological conditions of the samples. For example, \code{group = as.factor(c(rep("T", length(tumor)), rep("N", length(normal))))} specifies two groups, "T" and "N", where "T" represents tumor samples and "N" represents normal samples.}

\item{test}{The statistical test to be performed for the identification of the informative genes. If test="wilcoxon" a Wilcoxon test is performed. If test="var.test" an F-test on the variance is performed. The default is test="wilcoxon + var.test" and both Wilcoxon and var.test are performed.}

\item{percentile}{An integer indicating the percentile of the ranked distribution of informative genes to be returned. The default is percentile=10.}
}
\value{
An expression matrix for the informative genes to be used as input of the get_coexpression_network function.
}
\description{
The \code{get_informative_genes} function identifies informative genes, such as differentially expressed genes, that will be used to build a co-expressed network.
}
