% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacc_drug_targets_neighbors.R
\name{get_jacc_drug_targets_neighbors}
\alias{get_jacc_drug_targets_neighbors}
\title{Compute a squared matrix reporting the Jaccard similarity indices among couples of drug targets neighborhoods.}
\usage{
get_jacc_drug_targets_neighbors(Graph, target_nodes, neighborhood)
}
\arguments{
\item{Graph}{An igraph object.}

\item{target_nodes}{A vector of gene symbols likely to be drug targets.}

\item{neighborhood}{The order of neighborhood of the provided target nodes.}
}
\value{
A squared matrix reporting the Jaccard similarity indices among couples of drug targets neighborhoods.
}
\description{
Compute a squared matrix reporting the Jaccard similarity indices among couples of drug targets neighborhoods.
}
