% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modules.R
\name{get_modules}
\alias{get_modules}
\title{Get modules from the main igraph.}
\usage{
get_modules(iGraph, method = "walktrap")
}
\arguments{
\item{iGraph}{igraph object representing the main graph from which the subgraph should be extracted.}

\item{method}{community detection method from igraph.}
}
\value{
communities object cotaining communities identified from igraph object by using a specific method
}
\description{
Extract modules from the main graph by using a specified community detection algorithm from igraph.
}
