% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiple_coexpression_networks.R
\name{get_multiple_coexpression_networks}
\alias{get_multiple_coexpression_networks}
\title{Infer coexpression networks excluding the drugs for which less than minSamples samples are available}
\usage{
get_multiple_coexpression_networks(
  path,
  pattern = NULL,
  minSamples = 5,
  ncores = 12
)
}
\arguments{
\item{path}{The path where the expression matrices are stored.}

\item{pattern}{An optional regular expression. Only file names which match the regular expression will be returned.}

\item{minSamples}{Only the matrices with minSamples number of samples (columns) will be considered.}

\item{ncores}{Number of cores to infer the coexpression networks.}
}
\value{
A list of adjacency matrices.
}
\description{
Infer coexpression networks excluding the drugs for which less than minSamples samples are available
}
