% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ranked_gene_list.R
\name{get_ranked_gene_list}
\alias{get_ranked_gene_list}
\title{Get top ranked candidates computed by Borda on the provided list of attributes and the cutoff.}
\usage{
get_ranked_gene_list(iGraph, rank_list_attr, debug_output = FALSE)
}
\arguments{
\item{iGraph}{igraph object created from the binary consensus matrix computed by using get_ranked_consensus_binary_matrix() function.}

\item{rank_list_attr}{Vector of network attributes/scores to use for generating a combined gene rank}

\item{debug_output}{Print help and status messages to help debug the running of the function default:FALSE.}
}
\value{
vector of genes ordered by rank, based on the selected attributes associated to each gene.
}
\description{
get_ranked_gene_list uses the annotation associated with each vertex to rank them and generated a list of vertices names ordered by rank.
By default the annotations calculated by INfORM are "betweenness", "cc", "degree", "eccentricity", "closeness" & "eigenvector". The vertices are
ranked by each annotation separately nad then the ranks are unified by means of Borda(). User must provide the annotations to use in ranking scheme,
if the user has custom annotations such as "score" for differential gene expression then it can also be used for ranking.
}
