% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_smiles.R
\name{get_smiles_from_name}
\alias{get_smiles_from_name}
\title{Retrieves the CanonicalSmiles property for a compound from its name using the PubChem API.}
\usage{
get_smiles_from_name(compound)
}
\arguments{
\item{compound}{The name of the compound for which the CanonicalSmiles property should be retrieved.}
}
\value{
The CanonicalSmiles property for the compound, or \code{NA} if the retrieval fails.
}
\description{
This is an internal function that retrieves the CanonicalSmiles property for a given compound name using the PubChem API. It performs an HTTP GET request to the PubChem website and extracts the CanonicalSmiles property from the response.
}
