\name{getMatlabControl}
\alias{getMatlabControl}
\title{
    Get dream control settings from matlab .mat file
}    
\description{
  Get dream control settings from matlab .mat file
}
\usage{
getMatlabControl(mat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mat}{
    matlab .mat file
}
}
\value{
  list with same elements as dream control, see \code{\link{dream}}
}
