\name{getMatlabSeq}
\alias{getMatlabSeq}
\title{
  Get dream chains from matlab .mat file
}
\description{
  Read Reduced.Seq and convert to coda \code{\link{mcmc.list}}.
}
\usage{
getMatlabSeq(mat)
}
\arguments{
  \item{mat}{
    Matlab .mat file
}
}
\value{
\code{\link{mcmc.list}}
}
