% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{gof_rem}
\alias{gof_rem}
\title{Estimate the proportion of dyads predicted by \code{dream_rem} relational event model fits}
\usage{
gof_rem(object, realized = TRUE, rseed = NULL, ...)
}
\arguments{
\item{object}{An object of class "dream_rem".}

\item{realized}{How should cases be handled when the largest hazard for a specific event
is the same for the realized event and one or more control events? TRUE indicates that
the method will default to the realized event. FALSE indicates that the function will
take a sample from the set of realized and control events that have the largest event
hazard.}

\item{rseed}{The random seed for reproducibility of results. When more than 1 dyad has
the maximum hazard rate, the predicted dyad is selected at random with equal probabilty
for each dyad. For example, if three dyads have the same hazard value, then the probability
of selecting each dyad will be 1/3.}

\item{...}{Additional arguments for other methods.}
}
\value{
A \code{dream_gof} S3 object that contains the following results:
\itemize{
\item \code{props} - A \code{data.frame} that contains the proportion of correctly predicted dyads, senders, and targets.
\item \code{edgelist} - A \code{data.frame} object that contains the predicted and realized relational event sequence.
\item \code{estimated_rem} - The \code{dream_rem} S3 object provided to the function.
\item \code{match_realized} - The  \code{realized} argument provided to the function.
}
}
\description{
This function returns the proportion of dyads, senders, and targets predicted
by a \code{dream_rem} relational event model object. To compute the proportion for each
category, the function finds, for each realized event time, the dyad (based upon
the user's provided risk set) that is most likely to occur (i.e., the dyad
with the largest hazard rate) (Butts 2008). The predicted dyad is then compared to the
realized event dyad.
}
\details{
To compute the proportion for each category, the function finds, for each realized event time, the dyad (based upon
the user's provided risk set) that is most likely to occur (i.e., the dyad
with the largest hazard rate) (Butts 2008). The predicted dyad is then compared to the
realized event dyad.
}
\examples{
#Creating a psuedo one-mode relational event sequence with ordinal timing
relational.seq <- simulate_rem_seq(n_actors = 8,
                                  n_events = 50,
                                  inertia = TRUE,
                                  inertia_p = 0.10,
                                  sender_outdegree = TRUE,
                                  sender_outdegree_p = 0.05)

#Creating a post-processing event sequence for the above relational sequence
post.processing <-  create_res(type = "one-mode",
                              ordinal = TRUE,
                               riskset = "fixed",
                              time = relational.seq$eventID,
                              sender = as.character(relational.seq$sender),
                              receiver = as.character(relational.seq$target),
                              case_control=TRUE,
                              n_controls = 5)

#Computing the sender-outdegree statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_degree(formation = "sender-outdegree",
                                  data = post.processing,
                                  halflife = 2)

#Computing the inertia/repetition statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_repetition(data = post.processing,
                                        halflife = 2)

#Fitting an ordinal timing relational event model to the above one-mode relational
#event sequence
rem <- estimate_rem(~ sender.outdegree + repetition,
                         data=post.processing)
summary(rem) #summary of the relational event model

#the model goodness of fit
gof <- gof_rem(rem,rseed=3718)
summary(gof) #printing the summary of the goodness of fit assessment
plot(gof) #plotting the goodness of fit assessment
gof$edgelist #check the predicted edgelist for the dyadic events
}
\references{
Butts, Carter T. 2008. "A Relational Event Framework for Social Action." \emph{Sociological Methodology} 38(1): 155-200.
#'
}
