\name{LHSInit}
\alias{LHSInit}
\title{Latin Hypercube sampling...}
\usage{LHSInit(pars, nseq)}
\description{Latin Hypercube sampling}
\value{x matrix nseq x ndim. parameter-wise range [xmin, xmax]}
\arguments{\item{pars}{list. length ndim - one element per parameter}
\item{nseq}{scalar = number of sequences}}
