% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_opentargets_on_graph.R
\name{map_opentargets_on_graph}
\alias{map_opentargets_on_graph}
\title{Map OpenTargets taget drugs on a Gene Network}
\usage{
map_opentargets_on_graph(Graph, parsed_opentargets)
}
\arguments{
\item{Graph}{An igraph object.}

\item{parsed_opentargets}{A dataframe representing the OpenTargets database (e.g. obtained from the parse_opentargets_json).}
}
\value{
A dataframe reporting the drugs and diseases mapped on the provided graph.
}
\description{
This function maps OpenTargets target drugs onto the nodes of a given gene network.
}
