\name{maxLikPars}
\alias{maxLikPars}
\title{Naive maximum density parameter selection...}
\usage{maxLikPars(ss, ...)}
\description{Naive maximum density parameter selection}
\value{named character vector of parameter values

Uses which.max and density function}
\arguments{\item{ss}{MCMC chains. mcmc or mcmclist object}
\item{...}{extra parameters to pass to density}}
