% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_process.R
\name{new_dream_sequence}
\alias{new_dream_sequence}
\title{Constructor for \code{dream_sequence} objects}
\usage{
new_dream_sequence(
  x,
  statistics,
  ordinal,
  type,
  riskset,
  t,
  p,
  n,
  sampled_events,
  null,
  n_actors,
  n_senders,
  n_receivers,
  m,
  case_control,
  interevent_times
)
}
\arguments{
\item{x}{\code{data.frame} for the processed event sequence}

\item{statistics}{a \code{list} of the computed network statistics.}

\item{ordinal}{\code{logical} indicating if the sequence type is interval or ordinal timing.}

\item{type}{either of "one-mode" or "two-mode" characterizing the type of relational event sequence.}

\item{riskset}{\code{logical} indicating the definition of the risk set.}

\item{t}{If interval timing event sequence, the time point that marks the end of the sequence.}

\item{p}{The probability of sampling from the observed event sequence.}

\item{n}{The number of realized (observed) events.}

\item{sampled_events}{The number of sampled observed events.}

\item{null}{The number of non-realized (control) events.}

\item{n_actors}{The number of actors in the relational event sequence.}

\item{n_senders}{The number of event senders in the relational event sequence.}

\item{n_receivers}{The number of event receivers in the relational event sequence.}

\item{m}{The number of controls for each event.}

\item{case_control}{Boolean indicating if case-control sampling has occured.}

\item{interevent_times}{The numeric vector of interevent times (the timing between events).}
}
\description{
Constructor for \code{dream_sequence} objects
}
\keyword{internal}
