% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_edge_rank_matrix.R
\name{parse_edge_rank_matrix}
\alias{parse_edge_rank_matrix}
\title{This function parses an edge rank matrix and performs edge selection based on different strategies. It returns a list containing a binary matrix representing selected edges and a vector of edge ranks.}
\usage{
parse_edge_rank_matrix(
  edge_rank_matrix,
  edge_selection_strategy = "default",
  mat_weights = "rank",
  topN = 10,
  debug_output = FALSE,
  updateProgress = NULL
)
}
\arguments{
\item{edge_rank_matrix}{A symmetrix matrix with edge ranks as weight.}

\item{edge_selection_strategy}{Strategy to select edges for ensemble, "default" selects ranked edges untill all nodes have degree>=1, "top" swtiches to top N percentage ranked genes specfied by topN parameter.}

\item{mat_weights}{Type of scores in the ranked matrix; default:rank.}

\item{topN}{Top N percentage ranked edges to create ensemble if the edge_selection_strategy is "top"; default:10}

\item{debug_output}{Print help and status messages to help debug the running of the function; default:FALSE.}

\item{updateProgress}{Shiny application can request for update of progress from this function; default:NULL.}
}
\value{
A list containing a vector of consensus edge ranks and a binary symmetrix matrix representing the edge rank matrix.
}
\description{
This function parses an edge rank matrix and performs edge selection based on different strategies. It returns a list containing a binary matrix representing selected edges and a vector of edge ranks.
}
