% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_opentargets_json.R
\name{parse_opentargets_json}
\alias{parse_opentargets_json}
\title{Convenience function to parse OpenTargets JSON files}
\usage{
parse_opentargets_json(json_file)
}
\arguments{
\item{json_file}{The OpenTargets JSON file.}
}
\value{
A dataframe representing the OpenTargets database.
}
\description{
The parse_opentargets_json function is used to parse an OpenTargets JSON file and extract specific fields to create a final parsed data frame. The function takes the json_file as input and returns the parsed data frame.
}
