% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{plot.dream_gof}
\alias{plot.dream_gof}
\title{Plot method for \code{dream_gof} Relational Event Model Goodness-of-Fits}
\usage{
\method{plot}{dream_gof}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dream_gof}.}

\item{...}{Additional arguments for the barplot method.}
}
\description{
Creates barplots of a \code{dream_gof} to plot the proportion of
event dyads, senders, and receivers that are correctly specified.
}
\details{
Generate plots for \code{dream_gof} relational event model goodness-of-fits.
}
\examples{
#Creating a psuedo one-mode relational event sequence with ordinal timing
relational.seq <- simulate_rem_seq(n_actors = 8,
                                  n_events = 50,
                                  inertia = TRUE,
                                  inertia_p = 0.10,
                                  sender_outdegree = TRUE,
                                  sender_outdegree_p = 0.05)

#Creating a post-processing event sequence for the above relational sequence
post.processing <-  create_res(type = "one-mode",
                              ordinal = TRUE,
                               riskset = "fixed",
                              time = relational.seq$eventID,
                              sender = as.character(relational.seq$sender),
                              receiver = as.character(relational.seq$target),
                              case_control=TRUE,
                              n_controls = 5)

#Computing the sender-outdegree statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_degree(formation = "sender-outdegree",
                                  data = post.processing,
                                  halflife = 2)

#Computing the inertia/repetition statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_repetition(data = post.processing,
                                        halflife = 2)

#Fitting an ordinal timing relational event model to the above one-mode relational
#event sequence
rem <- estimate_rem(~ sender.outdegree + repetition,
                         data=post.processing)
summary(rem) #summary of the relational event model

#the model goodness of fit
gof <- gof_rem(rem,rseed=3718)
plot(gof) #plotting the goodness of fit assessment
}
