\name{possibility.envelope}
\alias{possibility.envelope}
\title{Obtain prediction confidence intervals for model, around new input values...}
\usage{possibility.envelope(dd, FUN.pars, ndraw=1000, conf=99)}
\description{Obtain prediction confidence intervals for model, around new input values}
\arguments{\item{dd}{dream object}
\item{FUN.pars}{new par values. if missing, use those from dream object}
\item{ndraw}{number of new iterations}
\item{conf}{Percentage two-sided confidence interval}}
