% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{predict.dream_rem}
\alias{predict.dream_rem}
\title{Predict method for Relational Event Model Fits}
\usage{
\method{predict}{dream_rem}(object, newdata = NULL, se.fit = FALSE, ...)
}
\arguments{
\item{object}{An object of class "dream_rem".}

\item{newdata}{If requested, a new model matrix (i.e., statistics) that the event rates will be based upon.}

\item{se.fit}{TRUE/FALSE. If TRUE, the standard errors of the predicted event rates will be returned.}

\item{...}{Additional arguments for other methods.}
}
\description{
Predicted event hazard rates based on \code{dream_rem} relational event model objects.
}
\details{
Following Butts (2008: 166), the rate for an event \eqn{a_i} at time \eqn{t} is formulated as:
\deqn{\lambda(a_i) = \exp[\lambda_0 + \theta^T z(s(a_i), r(a_i), x_i, A_t)]}
where \eqn{z()} is a mapping function that represents a set of sufficient
statistics for event \eqn{a_i}. \eqn{z()} represents the covariates included
in the estimated relational event model (i.e., the \code{object} argument) and
\eqn{\theta} are the estimated parameters/coefficients. \eqn{\lambda_0} is the
baseline rate across the relational event sequence and the intercept in the
interval relational event models and is arbitrary in the ordinal relational event
models. In addition, the standard error for each predicted value can be returned
when \code{se.fit} is set to TRUE, and is based upon the delta method, where the
standard error of the predicted values are:
\deqn{SE(\hat{\lambda}) = \sqrt(\lambda^2 diag(X\Sigma X^T))}
where \eqn{\hat{\lambda}} is the vector of predicted event rates, \eqn{X} is the
matrix of event covariates, and \eqn{\Sigma} is the covariance matrix of the
relational event model.
}
\examples{
#Creating a psuedo one-mode relational event sequence with ordinal timing
relational.seq <- simulate_rem_seq(n_actors = 8,
                                  n_events = 50,
                                  inertia = TRUE,
                                  inertia_p = 0.10,
                                  sender_outdegree = TRUE,
                                  sender_outdegree_p = 0.05)

#Creating a post-processing event sequence for the above relational sequence
post.processing <-  create_res(type = "one-mode",
                              ordinal = TRUE,
                               riskset = "constant_sample",
                              time = relational.seq$eventID,
                              sender = as.character(relational.seq$sender),
                              receiver = as.character(relational.seq$target),
                              n_controls = 5)

#Computing the sender-outdegree statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_degree(formation = "sender-outdegree",
                                  data = post.processing,
                                  halflife = 2)

#Computing the inertia/repetition statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_repetition(data = post.processing,
                                        halflife = 2)

#Fitting an ordinal timing relational event model to the above one-mode relational
#event sequence
rem <- estimate_rem(~ sender.outdegree + repetition,
                         data=post.processing)
summary(rem) #summary of the relational event model

#the predicted event rates
rates <- predict(rem)
hist(rates)
}
\references{
Butts, Carter T. 2008. "A Relational Event Framework for Social Action." \emph{Sociological Methodology} 38(1): 155-200.
}
