% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctionsDREAM.R
\name{print.dream}
\alias{print.dream}
\title{Print Method for Summary of dream Model}
\usage{
\method{print}{dream}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "summary.dream".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the main results of a 'dream' object.
}
\description{
Print Method for Summary of dream Model
}
