% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{print.dream_gof}
\alias{print.dream_gof}
\title{Print Method for \code{dream_gof} Model}
\usage{
\method{print}{dream_gof}(x, digits = 5, ...)
}
\arguments{
\item{x}{An object of class \code{dream_gof}.}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the main results of a 'dream_gof' object.
}
\description{
Print Method for \code{dream_gof} Model
}
