% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{print.dream_rem}
\alias{print.dream_rem}
\title{Print Method for dreamrem Model}
\usage{
\method{print}{dream_rem}(x, digits = 6, ...)
}
\arguments{
\item{x}{An object of class "dream_rem".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the main results of a 'dream' object.
}
\description{
Print Method for dreamrem Model
}
