% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_process.R
\name{print.dream_sequence}
\alias{print.dream_sequence}
\title{Print Method for 'dream'  object}
\usage{
\method{print}{dream_sequence}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class 'dream_support' .}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the main results of a 'dream' object.
}
\description{
Print Method for 'dream'  object
}
