% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctionsDREAM.R
\name{print.summary.dream}
\alias{print.summary.dream}
\title{Print Method for dream Model}
\usage{
\method{print}{summary.dream}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "dream".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the main results of a 'dream' summary object.
}
\description{
Print Method for dream Model
}
