% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{print.summary.dream_gof}
\alias{print.summary.dream_gof}
\title{Print Method for \code{summary.dream_gof} Model}
\usage{
\method{print}{summary.dream_gof}(x, digits = 5, ...)
}
\arguments{
\item{x}{An object of class \code{summary.dream_gof}}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the main results of a \code{summary.dream_gof} summary object.
}
\description{
Print Method for \code{summary.dream_gof} Model
}
