% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_process.R
\name{print.summary.dream_sequence}
\alias{print.summary.dream_sequence}
\title{Print Method for dream Model}
\usage{
\method{print}{summary.dream_sequence}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class "dream_sequence".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the main results of a 'dream_sequence' summary object.
}
\description{
Print Method for dream Model
}
