% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tac.R
\name{query_tac}
\alias{query_tac}
\title{Query NOAA Tides and Currents stations}
\usage{
query_tac(
  start,
  end,
  product,
  station,
  step = "3 days",
  datum = "MLLW",
  tz = "gmt"
)
}
\arguments{
\item{start}{Character vector start date for query ("YYYY-MM-DD").}

\item{end}{Character vector end date for query ("YYYY-MM-DD").}

\item{product}{Specify data product type. Defaults to "water_temperature".
See a complete listing of available products at \url{https://tidesandcurrents.noaa.gov/api/#products}. Note that all products are
not available at all stations. All data are returned in metric units.}

\item{station}{Station ID number. A complete listing of stations is available at \url{https://tidesandcurrents.noaa.gov/stations.html}.}

\item{step}{Time step for data query. Default time step is 3 days ("3 days"), with a maximum time step of 30 days.
Larger time steps will result in faster downloads for longer time periods.}

\item{datum}{Queries regarding water level require a datum. If none is selected, then datum defaults to "MLLW," or Mean-Lower Low Water. Example datum selections are available to view at \url{https://tidesandcurrents.noaa.gov/datums.html?id=8722607}.}

\item{tz}{Data product time zone. Defaults to GMT ("gmt"). Other options include Local Standard Time ("lst") and Local Standard/Local Daylight Time ("lst_ldt").}
}
\description{
Query NOAA Tides and Currents stations
}
\examples{
query_tac(start = "1994-03-16",
          end ="1994-05-18",
          step = "30 days",
          product = "water_temperature",
          station = 8631044)
}
