% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kml.R
\name{read_kml}
\alias{read_kml}
\title{Read KML/KMZ files}
\usage{
read_kml(
  filename,
  cast = "LINESTRING",
  crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{filename}{Path to .kml or .kmz file}

\item{crs}{Coordinate reference system. Defaults to WGS84.}
}
\value{
Returns an sf object.
}
\description{
A function to read KML/KMZ files.
}
\examples{
fname <- system.file("extdata/MAB_poly.kml", package = "dream")
read_kml(fname)

}
