% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{residuals.dream_rem}
\alias{residuals.dream_rem}
\title{Model residuals for \code{dream_rem} relational event model fits}
\usage{
\method{residuals}{dream_rem}(object, ...)
}
\arguments{
\item{object}{An object of class "dream_rem".}

\item{...}{Additional arguments for other methods.}
}
\value{
A \code{data.frame} object that contains the following results:
\itemize{
\item \code{time} - The timing of each realized event.
\item \code{unit.deviance} - The unit deviance for each realized event (the contribution of each realized event to the log-likelihood).
\item \code{residual.deviance} - The residual deviance for each realized event (the square of the contribution of each realized event to the log-likelihood).
\item \code{standardized.deviance} - The standardized residual deviance for each realized event (the square of the contribution of each realized event to the log-likelihood).
\item \code{schof.resid} - The Schoenfeld residual for each included covariate.
}
}
\description{
This function returns a set of model residuals based upon the realized
events in a relational event sequence (Butts 2008). The residuals included are: unit deviance,
residual deviance, standardized residual deviance, and for each covariate,
the Schoenfeld residual (please see Schoenfeld 1982).
}
\details{
The residuals, based upon an estimated relational event model, included are: unit deviance,
residual deviance, standardized residual deviance, and for each covariate,
the Schoenfeld residual (please see Schoenfeld 1982).
}
\examples{
#Creating a psuedo one-mode relational event sequence with ordinal timing
relational.seq <- simulate_rem_seq(n_actors = 8,
                                  n_events = 50,
                                  inertia = TRUE,
                                  inertia_p = 0.10,
                                  sender_outdegree = TRUE,
                                  sender_outdegree_p = 0.05)

#Creating a post-processing event sequence for the above relational sequence
post.processing <-  create_res(type = "one-mode",
                              ordinal = TRUE,
                               riskset = "constant_sample",
                              time = relational.seq$eventID,
                              sender = as.character(relational.seq$sender),
                              receiver = as.character(relational.seq$target),
                              n_controls = 5)

#Computing the sender-outdegree statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_degree(formation = "sender-outdegree",
                                  data = post.processing,
                                  halflife = 2)

#Computing the inertia/repetition statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_repetition(data = post.processing,
                                        halflife = 2)

#Fitting an ordinal timing relational event model to the above one-mode relational
#event sequence
rem <- estimate_rem(~ sender.outdegree + repetition,
                         data=post.processing)
summary(rem) #summary of the relational event model

#the model residuals
residuals(rem)
}
\references{
Butts, Carter T. 2008. "A Relational Event Framework for Social Action." \emph{Sociological Methodology} 38(1): 155-200.

Schoenfeld, David. 1982 "Partial Residuals for The Proportional Hazards Regression Model." \emph{Biometrika} 69(1): 239-241.
}
