% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rst_to_tib.R
\name{rst_to_tib}
\alias{rst_to_tib}
\title{Convert rasters to tibbles.}
\usage{
rst_to_tib(r, var_name = "fill_var", xy_name = c("longitude", "latitude"))
}
\arguments{
\item{r}{Raster object to be converted.}

\item{var_name}{Column name for fill variable. Defaults to \code{fill_var}.}

\item{xy_name}{Column names for x and y variables (in that order). Defaults to \code{longitude} and \code{latitude}.}
}
\value{
A \code{tibble} with three columns: a fill variable, longitude, and latitude.
}
\description{
Convert rasters to tibbles.
}
\examples{

library(raster)
f <- system.file("external/test.grd", package="raster")
r <- raster(f)
tib <- rst_to_tib(r)
tib
}
