% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Convenience function to convert sfc to lat/lon columns within a data.frame.}
\usage{
sfc_as_cols(x, geometry, names = c("x", "y"))
}
\arguments{
\item{x}{An \code{sf} object}

\item{names}{Names for longitude and latitude columns}
}
\description{
Convenience function to convert sfc to lat/lon columns within a data.frame.
}
\details{
Originally written by Josh London here -> https://github.com/r-spatial/sf/issues/231
}
