% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_data.R
\docType{data}
\name{southern.women}
\alias{southern.women}
\title{Davis Southern Women's Dataset}
\format{
\subsection{\code{southern.women}}{

Two-Mode affliation matrix from Davis et al.(1941) Southern Women study. 18 women x 14 events.
Dataset is taken from the networkdata R package (Almquist 2014)
}
}
\source{
Almquist, Zach. 2014. \emph{networkdata: Lin Freeman's Network Data Collection}. R package version 0.01, \url{https://github.com/Z-co/networkdata}.

Brieger, Ronald. 1974. "Duality of Persons and Groups." \emph{Social Forces} 53(2): 181-190.

Davis, Allison, Burleigh B. Gardner, and Mary R. Gardner. 1941. \emph{Deep South: A Social Anthropological Study of Caste and Class}. University of Chicago Press.
}
\usage{
data(southern.women)
}
\description{
Davis Southern Women's Dataset
}
\keyword{datasets}
