% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_management.R
\name{start_server}
\alias{start_server}
\title{Start the Flask API server}
\usage{
start_server(python = "python")
}
\arguments{
\item{python}{Path to the Python interpreter (default is "python")}
}
\value{
An instance of the "process" class representing the server process
}
\description{
This function starts a server using the specified Python interpreter.
It creates a new process and runs the "dream.api" module using the "-m" flag.
The standard output and standard error of the process are redirected to two log files.
}
