% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_management.R
\name{stop_server}
\alias{stop_server}
\title{Stop the Flask API server}
\usage{
stop_server(p)
}
\arguments{
\item{p}{An instance of the "process" class representing the server process}
}
\description{
This function stops the server process. It sends an interrupt signal to the
process to gracefully terminate it.
}
