% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctionsDREAM.R
\name{summary.dream}
\alias{summary.dream}
\title{Summary Method for dream Objects}
\usage{
\method{summary}{dream}(object, digits = 4, ...)
}
\arguments{
\item{object}{An object of class "dream".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of summary statistics for the relational event model
including parameter estimates, (null) likelihoods, and tests of significance
for likelihood ratios and estimated parameters.
}
\description{
Summarizes the results of an ordinal timing relational event model.
}
