% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{summary.dream_gof}
\alias{summary.dream_gof}
\title{Summary Method for \code{dream_gof} Objects}
\usage{
\method{summary}{dream_gof}(object, digits = 5, ...)
}
\arguments{
\item{object}{An object of class \code{dream_gof}.}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of summary statistics for the goodness-of-fit assessment.
}
\description{
Summarizes the results of a goodness-of-fit assessment for
estimated relational event models.
}
