% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{summary.dream_rem}
\alias{summary.dream_rem}
\title{Summary Method for dreamrem Objects}
\usage{
\method{summary}{dream_rem}(object, digits = 6, ...)
}
\arguments{
\item{object}{An object of class "dream_rem".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of summary statistics for the relational event model
including parameter estimates, (null) likelihoods, and tests of significance
for likelihood ratios and estimated parameters.
}
\description{
Summarizes the results of an ordinal timing relational event model.
}
