% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_process.R
\name{summary.dream_sequence}
\alias{summary.dream_sequence}
\title{Summary Method for dream_sequence Objects}
\usage{
\method{summary}{dream_sequence}(object, digits = 4, ...)
}
\arguments{
\item{object}{An object of class "dream_sequence".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of descripitive statistics for a processed relational event sequence.
}
\description{
Summarizes the main components of a processed relational event sequence.
}
