% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_similarity_significance.R
\name{test_similarity_significance}
\alias{test_similarity_significance}
\title{Performs a permutation test to assess the significance of the jaccard scores among couples of drug targets (e.g. obtained from the get_jacc_drug_targets_neighbors function).}
\usage{
test_similarity_significance(drug_target_mat, Graph, nIter = 1000)
}
\arguments{
\item{drug_target_mat}{A squared matrix containing similarity scores between drug targets pairs (e.g. obtained from the get_jacc_drug_targets_neighbors function).}

\item{Graph}{An igraph object.}

\item{nIter}{Number of permutations to be performed.}
}
\value{
A squared matrix reporting the statistical significance of the similarity measures (e.g. Jaccard indices) among couples of drug targets neighborhoods.
}
\description{
Performs a permutation test to assess the significance of the jaccard scores among couples of drug targets (e.g. obtained from the get_jacc_drug_targets_neighbors function).
}
