% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fade.R
\name{theme_fade}
\alias{theme_fade}
\title{A ggplot2 theme}
\usage{
theme_fade(...)
}
\description{
A ggplot2 theme
}
\examples{

library(ggplot2)
ggplot(iris) +
   geom_point(aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
   theme_fade()
}
