% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Assess model fit for multiple models through DHARMa residual simulation}
\usage{
validate(mod_set, n = 1000)
}
\arguments{
\item{mod_set}{A list of model objects.}

\item{n}{Number of DHARMa simulations to generate from each model.}
}
\value{
Returns a data.frame containing model formulas for each model and P values for
each test applied.
}
\description{
This function applies DHARMa validation tests to a user-specified list of models. Tests
look for agreement between simulated and empirical residuals; assessing models for dispersion,
zero-inflation, anomalous prevalence of outliers, and deviations from expected distributions.
}
\details{
Applies DHARMa residual diagnostics test to a list of models.
}
\examples{

library(glmmTMB)
mod_list <- list()
mod_list[[1]] <- glmmTMB(count ~ mined + (1|site), zi=~mined, family=poisson, data=Salamanders)
mod_list[[2]] <- glmmTMB(count ~ mined + cover + (1|site), zi=~mined, family=poisson, data=Salamanders)
mod_list[[3]] <- glmmTMB(count ~ mined + cover + spp + (1|site), zi=~mined, family=poisson, data=Salamanders)
dream::validate(mod_list)
}
\references{
Hartig, Florian. 2017. “DHARMa: Residual Diagnostics for Hierarchical (Multi-Level/Mixed) Regression Models.” R Package Version 0.1 5.
}
