% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_process.R
\name{validate_dream_sequence}
\alias{validate_dream_sequence}
\title{Validator for \code{dream_sequence} objects}
\usage{
validate_dream_sequence(x)
}
\arguments{
\item{x}{a created \code{dream_sequence} object to be checked.}
}
\description{
Validator for \code{dream_sequence} objects
}
\keyword{internal}
