% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_table.R
\name{dream_table}
\alias{dream_table}
\title{Generate pretty LaTeX tables from glmmTMB summaries}
\usage{
dream_table(model, caption)
}
\arguments{
\item{model}{A glmmTMB model object.}

\item{caption}{A caption to include with the table.}
}
\description{
Generate pretty LaTeX tables from glmmTMB summaries
}
\examples{
data(cbpp, package="lme4")
bovine <- glmmTMB(cbind(incidence, size-incidence) ~ period + (1|herd),
                 family=binomial, data=cbpp)
dream_table(model = bovine, caption = "Dream table for Binomial GLMM")
}
