% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_remstats.R
\name{dreamstats_dyadfe}
\alias{dreamstats_dyadfe}
\title{Add Dyadic-Level Fixed Effects for Event Dyads in a Relational Event Sequence}
\usage{
dreamstats_dyadfe(data, directed = TRUE, return_stats = FALSE)
}
\arguments{
\item{data}{A \code{dream_sequence} object that contains the processed relational event sequence.}

\item{directed}{TRUE/FALSE. TRUE indicates that the dyadic-level fixed effects
will be generated based upon the ordering of the sending and receiving actors (i.e.,
AB != BA). FALSE indicates that the dyadic-level fixed effects will be generated
based upon the combo of the sending and receiving actors (i.e., AB == BA). Set to
TRUE by default. Of course, these will lead to (1) a different number of fixed
effects and (2) a different interpretation of the results.}

\item{return_stats}{TRUE/FALSE. TRUE indicates that the vector of computed
statistics will be returned. FALSE indicates that the vector of computed
statistics will be added to the \code{statistics} element of the \code{data} argument.}
}
\value{
The vector of actor-level fixed effects for the relational event sequence or the updated \code{data} argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function allows users to add dyad-level fixed effects
for relational event models to a \code{dream_sequence} object.
}
\details{
This function adds dyad-level fixed effects to a \code{dream_sequence}
object. Internally, the function creates a new variable named "dyad.fe"
which is a factor of the combined event sender/receiver ids.
}
\examples{
events <- data.frame(time = 1:18, eventID = 1:18,
                    sender = c("A", "B", "C",
                               "A", "D", "E",
                               "F", "B", "A",
                               "F", "D", "B",
                               "G", "B", "D",
                               "H", "A", "D"),
                    target = c("B", "C", "D",
                               "E", "A", "F",
                              "D", "A", "C",
                               "G", "B", "C",
                               "H", "J", "A",
                               "F", "C", "B"))

processed <- create_res(type = "one-mode",
                      ordinal = TRUE,
                      riskset = "complete",
                      time = events$time,
                      sender = events$sender,
                      receiver = events$target,
                      seed = 9999)

#adding the dyadic fixed effects to the statistics list
processed <- dreamstats_dyadfe(data=processed)

#estimating the dyadic fixed effects only ordinal timing relational event model
model <- estimate_rem(~dyad.fe, data = processed)

}
\author{
Kevin A. Carson \href{mailto:kacarson@arizona.edu}{kacarson@arizona.edu}, Diego F. Leal \href{mailto:dflc@arizona.edu}{dflc@arizona.edu}
}
