% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_remstats.R
\name{dreamstats_dyadic}
\alias{dreamstats_dyadic}
\title{Add Dyadic-Level Statistics for Event Dyads in a Relational Event Sequence}
\usage{
dreamstats_dyadic(
  data,
  var_name,
  transformation = c("same", "abs.diff", "inv.abs.diff"),
  return_stats = FALSE
)
}
\arguments{
\item{data}{A \code{dream_sequence} object that contains the processed relational event sequence.}

\item{var_name}{A string that is the name of the variable from the \code{statistics} element in the \code{dream_sequence} argument for which the statistic should be created.}

\item{transformation}{The type of transformation for how the sender and receiver values will be compared (see details). The following arguments are provided: "same", "abs.diff", and "inv.abs.diff".}

\item{return_stats}{TRUE/FALSE. TRUE indicates that the vector of computed
statistics will be returned. FALSE indicates that the vector of computed
statistics will be added to the \code{statistics} element of the \code{data} argument.}
}
\value{
The vector of dyadic-level statistics for the relational event sequence or the updated \code{data} argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function allows users to add time-varying and time-invariant
dyadic-level statistics that impact the dyadic event rates in
relational event models.
}
\details{
This function adds user-provided time-varying and time-invariant dyadic-level
statistics to a \code{dream_sequence} object for relational event models. For this
function to work, in the object provided to the \code{data} argument, the \code{statistics}
element must contain the following previously computed statistics: \code{var_name.sender} and
\code{var_name.receiver}, where \code{var_name} is the user-provided argument. For instance,
if the \code{var_name} is \code{male}, then the two variables need to be included in the
\code{statistics} list: \code{male.sender} and \code{male.receiver}.

The function allows for three types of transformations to compare the values
for the event senders and event receivers. When the \code{transformation} argument
is \code{same}, then the values are 1 if the event sender and receivers are the
same, and 0 if not. When the \code{transformation} argument
is \code{abs.diff}, then the values are \eqn{|sender - receiver|}. Finally, when
the \code{transformation} argument is \code{inv.abs.diff}, then the values are \eqn{1/|sender - receiver|} (if the difference is 0, the value is set to 1).
}
\examples{
events <- data.frame(time = 1:18, eventID = 1:18,
                    sender = c("A", "B", "C",
                               "A", "D", "E",
                               "F", "B", "A",
                               "F", "D", "B",
                               "G", "B", "D",
                               "H", "A", "D"),
                    target = c("B", "C", "D",
                               "E", "A", "F",
                              "D", "A", "C",
                               "G", "B", "C",
                               "H", "J", "A",
                               "F", "C", "B"))

processed <- create_res(type = "one-mode",
                      ordinal = TRUE,
                      riskset = "constant_sample",
                      time = events$time,
                      sender = events$sender,
                      receiver = events$target,
                      p_samplingobserved = 1.00,
                      n_controls = 1,
                      seed = 9999)

#reconstructing the data.frame object to store the time-varying
#actor-level statistic for receivers
ids <- unique(c(events$sender,events$target))
actor_stats <- data.frame(actor_id = ids,
                         time_start = 0,
                         time_end = 18,
                         male = sample(0:1, length(ids), TRUE))

#adding the value to the post-processing relational event sequence where
#the new variable is named "male"
processed <- dreamstats_actor(data = processed,
                             var_name = "male",
                             sender_ref = TRUE,
                             actor_info = actor_stats)
processed
processed$statistics$male.sender
#adding the value for the event receivers
processed <- dreamstats_actor(data = processed,
                             var_name = "male",
                             sender_ref = FALSE,
                             actor_info = actor_stats)
processed
processed$statistics$male.receiver


#adding the dyadic same value to the post-processing relational event sequence where
#the new variable is 1 if the event sender and receiver have the same value
# and 0 if not
processed <- dreamstats_dyadic(data = processed,
                             var_name = "male",
                             transformation = "same")
processed
processed$statistics$male.same

extract.data <- as.data.frame(processed)
extract.data #checking the values are the same!

}
\author{
Kevin A. Carson \href{mailto:kacarson@arizona.edu}{kacarson@arizona.edu}, Diego F. Leal \href{mailto:dflc@arizona.edu}{dflc@arizona.edu}
}
