% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_remstats.R
\name{dreamstats_event}
\alias{dreamstats_event}
\title{Add Event-Level Statistics for a Relational Event Sequence}
\usage{
dreamstats_event(
  data,
  var_name,
  event_info,
  make_factor = FALSE,
  return_stats = FALSE
)
}
\arguments{
\item{data}{A \code{dream_sequence} object that contains the processed relational event sequence.}

\item{var_name}{A string that is the name of the variable from the \code{event_info} argument that represents the actor-level statistic to be added.}

\item{event_info}{A \eqn{N} x 2 \code{data.frame} object that contains two columns with the number of observations
being the number of observed time points actors. The object should contain
two named columns: (1) \code{time_id}, which is the time associated with each event and (2) \code{var_name} (from the argument), where
the var_name column is the vector of statistics (each time point has one associated value; see details for more information).}

\item{make_factor}{TRUE/FALSE. TRUE indicates that the vector of values will be made a factor, and FALSE if not.}

\item{return_stats}{TRUE/FALSE. TRUE indicates that the vector of computed
statistics will be returned. FALSE indicates that the vector of computed
statistics will be added to the \code{statistics} element of the \code{data} argument.}
}
\value{
The vector of event-level statistics for the relational event sequence or the updated \code{data} argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function allows users to add event-level statistics that impact the event rates in
interval timing relational event model, such as statistics that impact the
waiting times between events.
}
\details{
This function adds user-provided time-varying and time-invariant actor-level
statistics to a \code{dream_sequence} object for relational event models. The \code{event_info}
argument should be a \eqn{N} x 2 \code{data.frame} object with two named columns. The first
column should be named \code{time_id}, which represents the observed time points
based upon the relational event sequence. The second column should be named after the
\code{var_name} argument and represents the event-level statistic for that i time point (i.e., the ith \code{time_id}).
}
\examples{
events <- data.frame(time = 1:18, eventID = 1:18,
                    sender = c("A", "B", "C",
                               "A", "D", "E",
                               "F", "B", "A",
                               "F", "D", "B",
                               "G", "B", "D",
                               "H", "A", "D"),
                    target = c("B", "C", "D",
                               "E", "A", "F",
                              "D", "A", "C",
                               "G", "B", "C",
                               "H", "J", "A",
                               "F", "C", "B"))

processed <- create_res(type = "one-mode",
                      ordinal = TRUE,
                      time = events$time,
                      riskset = "constant_sample",
                      sender = events$sender,
                      receiver = events$target,
                      p_samplingobserved = 1.00,
                      n_controls = 5,
                      seed = 9999)

#reconstructing the data.frame object to store the time-varying
#event-level statistic
event_stats <- data.frame(time_id = events$time,
                         oliver = rnorm(nrow(events)))

#reconstructing the data.frame object to store the time-varying
#event-level statistic
processed <- dreamstats_event(data = processed,
                             var_name = "oliver",
                             event_info = event_stats)
processed
processed$statistics$oliver.event

extract.data <- as.data.frame(processed)
extract.data

}
\author{
Kevin A. Carson \href{mailto:kacarson@arizona.edu}{kacarson@arizona.edu}, Diego F. Leal \href{mailto:dflc@arizona.edu}{dflc@arizona.edu}
}
