% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{estimate_rem}
\alias{estimate_rem}
\title{Fit a Maximum Likelihood Relational Event Model (REM) to A Processed Relational Event Sequence}
\usage{
estimate_rem(
  formula,
  data,
  newton.rhapson = TRUE,
  optim.method = "BFGS",
  optim.control = list(),
  tolerance = 1e-09,
  maxit = 100,
  starting.beta = NULL,
  multiple.events = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula object where the covariates are on the right hand
side of ~. The names of the covariates must follow the names in the
\code{statistics} element of the \code{data} argument, since the right hand side covariates
are taken from this list. The dependent variable does not need to be defined, since it is internally
defined based upon the \code{data} argument. This is the same argument found in \code{\link[stats]{lm}} and \code{\link[stats]{glm}}.}

\item{data}{An object of class \code{dream_sequence} that contains the processed relational event sequence and
statistics that are included in the \code{formula} argument.}

\item{newton.rhapson}{TRUE/FALSE. TRUE indicates an internal Newton-Rhapson iteration procedure with line searching is used to
find the set of maximum likelihood estimates. FALSE indicates that the log likelihood function will be optimized via the
\code{\link{optim}} function. The function defaults to TRUE.}

\item{optim.method}{If newton.rhapson is FALSE, what optim method should be used in conjunction with the \code{\link{optim}} function. Defaults
to "BFGS". See the \code{\link{optim}} function for the set of options.}

\item{optim.control}{If newton.rhapson is FALSE, a list of control to be used in the \code{\link{optim}} function. See the \code{\link{optim}} function for the set of controls.}

\item{tolerance}{If newton.rhapson is TRUE, the stopping criterion for the absolute difference in the log likelihoods for each Newton-Rhapson iteration.
The optimization procedure stops when the absolute change in the log likelihoods is less than \code{tolerance} (see Greene 2003).}

\item{maxit}{If newton.rhapson is TRUE, the maximum number of iterations for the Newton-Rhapson optimization procedure (see Greene 2003).}

\item{starting.beta}{A numeric vector that represents the starting parameter estimates for the Newton-Rhapson optimization procedure. This may be a beneficial argument
if the optimization procedure fails, since the Newton-Rhapson optimization procedure is sensitive to starting values. Preset to NULL.}

\item{multiple.events}{TRUE/FALSE. Currently, this function assumes that only one event occurs per
event cluster (i.e., time point). Future versions of the package will include estimation options for
multiple events per time point, commonly referred to as tied events, via the Breslow approximation
technique (see Box-Steffensmeier and Jones 2004). At this moment, this argument is preset to FALSE
and should not be modified by the user.}

\item{...}{Additional arguments.}
}
\value{
An object of class "dream_rem" as a list containing the following components:
\itemize{
\item \code{optimization.method} - The optimzation method used to find the parameters.
\item \code{converged} - TRUE/FALSE. TRUE indicates that the REM converged.
\item \code{loglikelihood.null} - The log likelihood of the null model (i.e., the model where the parameters are assumed to be 0).
\item \code{loglikelihood.full} - The log likelihood of the estimated model.
\item \code{chi.stat} - The chi-statistic of the likelihood ratio test.
\item \code{loglikelihood.test} - The p-value of the likelihood ratio test.
\item \code{df.null} - The degrees of freedom of the null model.
\item \code{df.full} - The degrees of freedom of the full model.
\item \code{parameters} - The MLE parameter estimates.
\item \code{hessian} - The estimated hessian matrix.
\item \code{gradient} - The estimated gradient vector.
\item \code{se.parameter} - The standard errors of the MLE parameter estimates.
\item \code{covariance.mat} - The estimated variance-covariance matrix.
\item \code{z.values} - The z-scores for the MLE parameter estimates.
\item \code{p.values} - The p-values for the MLE parameter estimates.
\item \code{AIC} - The AIC of the estimated REM.
\item \code{BIC} - The BIC of the estimated REM.
\item \code{n.events} - The number of observed events in the relational event sequence.
\item \code{null.events} - The number of control events in the relational event sequence.
\item \code{newton.iterations} - The number of Newton-Rhapson iterations.
\item \code{search.algo} - A data.frame object that contains the Newton-Rhapson searching algorithm results.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function estimates the ordinal and interval timing relational event model by maximizing the
likelihood function given by Butts (2008) via maximum likelihood estimation. A nice outcome
is that the ordinal timing relational event model is equivalent to the conditional logistic
regression (see Greene 2003; for R functions, see \code{\link[survival]{clogit}}). In
addition, based on this outcome and the structure of the data, this function can estimate
the Cox proportional hazards model (see Box-Steffensmeier and Jones 2004; for R functions, see \code{\link[survival]{coxph}})
given that the likelihood functions are equivalent. An important assumption this model
makes is that only one event occurs at each time point. If this is unfeasible for
the user's specific dataset, we encourage the user to see the \code{\link[survival]{clogit}}
function for the Breslow approximation technique (Box-Steffensmeier and Jones 2004). Future
versions of the package will include options for tied event data (e.g., multiple events at one time point).
}
\details{
This function estimates the ordinal and interval timing relational event model by maximizing the
likelihood function provided in the seminal REM paper by Butts (2008) via maximum likelihood estimation. The
ordinal timing likelihood function is:
\deqn{L(A_t|\beta) = \prod_{i=1}^{|A_t|} \frac{\lambda_{a_i}}{\sum_{a' \in M_t \lambda_{a'}}}}
where, following Butts (2008) and Duxbury (2020), \eqn{A_t} is the relational event sequence,
\eqn{\lambda_{a_i}} is the hazard rate for event \emph{i}, which is formulated to be equal to
\eqn{exp(\beta^{T}z(x,Y))}, that is, the linear combination of user-specific covariates, \eqn{z(x,Y)}, and associated
REM parameters, \eqn{\beta}. The user provides these covariates via the \code{formula} argument. \eqn{M_t} is the support set for event \eqn{a_i \in A_t}. The likelihood function for the interval timing relational event
model is:
\deqn{L(A_t | \beta) = [ \prod_{i=1}^{|A_t|} \lambda_i \prod_{j \in M_{\tau(i)}} \exp(-\lambda_j \{\tau(i) - \tau(i-1) \}) ] \times [\prod_{j \in M_t} \exp(-\lambda_j \{t - \tau(M) \})]}
where \eqn{\tau(i)} is the time of the observed (realized) event \emph{i} and \eqn{t} is the time that marks the end of the relational event sequence.
Following Duxbury (2020), \eqn{z(x,Y)} is a mapping function that represents the endogenous network statistics computed on the network
of past events,\eqn{x}, and exogenous covariates, \eqn{Y}. In comparison to the ordinal
timing relational event formulation, the hazard rate for event \emph{i}, \eqn{\lambda_{a_i}}, includes
the baseline hazard rate (the intercept), \eqn{exp(\beta_0 + \beta^{T}z(x,Y))}. If \eqn{t} is not known
by the user, then the interval timing likelihood is:
\deqn{L(A_t | \beta) =  \prod_{i=1}^{|A_t|} \lambda_i \prod_{j \in M_{\tau(i)}} \exp(-\lambda_j \{\tau(i) - \tau(i-1) \})  }
In this case, the likelihood function is the same as employed in the \code{\link[remstimate]{remstimate}}
for interval timing relational event models. The values for \eqn{t} are taken from the
\code{data} object.

This function provides two numerical optimization techniques to find the maximum
likelihood estimates for the associated parameters. First, this function allows
the user to use the \code{\link{optim}} function to find the associated parameters
based on the above likelihood function. Secondly, and by default, this function
employs a Newton-Rhapson iteration algorithm with line-searching to find
the unknown parameters (see Greene 2003 for a discussion of this algorithm). If desired, the user can
provide the initial searching values for both algorithms with the \code{starting.beta} argument.

It's important to note that the modeling concerns of the conditional logistic regression apply to the
ordinal timing relational event model, such as no within-sequence fixed effects, that is,
a variable that does not vary within event cluster (i.e., a variable that is the
same for both the null and observed events). The function internally checks for
this and provides the user with a warning if any requested effects has no total within-event
variance. Moreover, any observed events that have no associated control events are
removed from the analysis as they provide no information to the log likelihood (see Greene 2003). The function
removes these events from the sequence prior to estimation.
}
\examples{
#Creating a psuedo one-mode relational event sequence with ordinal timing
relational.seq <- simulate_rem_seq(n_actors = 8,
                                  n_events = 50,
                                  inertia = TRUE,
                                  inertia_p = 0.10,
                                  sender_outdegree = TRUE,
                                  sender_outdegree_p = 0.05)

#Creating a post-processing event sequence for the above relational sequence
post.processing <-  create_res(type = "one-mode",
                              ordinal = TRUE,
                              riskset = "complete",
                              time = relational.seq$eventID,
                              sender = as.character(relational.seq$sender),
                              receiver = as.character(relational.seq$target))

#Computing the sender-outdegree statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_degree(formation = "sender-outdegree",
                                  data = post.processing,
                                  halflife = 2)

#Computing the inertia/repetition statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_repetition(data = post.processing,
                                      halflife = 2)

#Fitting an ordinal timing relational event model to the above one-mode relational
#event sequence
rem <- estimate_rem(~ sender.outdegree + repetition,
                         data=post.processing)
summary(rem) #summary of the relational event model

vcov(rem) #printing the variance-covariance matrix
logLik(rem) #printing the model log-likelihood
AIC(rem) #printing the model AIC
rates <- predict(rem) #extracting the predicted event rates


#Fitting a (ordinal) relational event model to the above one-mode relational
#event sequence via the optim function
rem1 <- estimate_rem(~ sender.outdegree + repetition,
                   data=post.processing,
                   newton.rhapson=FALSE)
summary(rem1) #summary of the relational event model


# a psuedo relational event sequence
events <- data.frame(time = 1:18, eventID = 1:18,
                                sender = c("A", "B", "C",
                                           "A", "D", "E",
                                           "F", "B", "A",
                                           "F", "D", "B",
                                           "G", "B", "D",
                                          "H", "A", "D"),
                               target = c("B", "C", "D",
                                          "E", "A", "F",
                                          "D", "A", "C",
                                          "G", "B", "C",
                                          "H", "J", "A",
                                          "F", "C", "B"))

# Creating a dynamic one-mode relational risk set with p = 1.00 (all true events)
# and 5 controls based upon the interval timing relational event framework
eventSet <- create_res(ordinal = FALSE,
                      t = max(events$time) + rexp(1),
                      riskset = "dynamic_sample",
                      type = "one-mode",
                      time = events$time,
                      sender = events$sender,
                      receiver = events$target,
                      p_samplingobserved = 1.00,
                      n_controls = 5,
                      seed = 9999)

#Computing the sender indegree statistic for the relational event sequence
eventSet <- dreamstats_degree(formation = "sender-indegree",
                             data = eventSet,
                             halflife = 2)

#Computing the outgoing two path statistic for the relational event sequence
eventSet <- dreamstats_triads(formation = "OTP",
                             data = eventSet,
                             halflife = 2)

#Fitting an interval timing relational event model to the above one-mode relational
#event sequence
rem.interval <- estimate_rem(~ sender.indegree + outgoing.two.paths,
                            data=eventSet)
summary(rem.interval) #summary of the relational event model

rem.interval.optim <- estimate_rem(~ sender.indegree + outgoing.two.paths,
                                  data=eventSet,
                                  newton.rhapson=FALSE)
summary(rem.interval.optim) #summary of the relational event model
}
\references{
Box-Steffensmeier, Janet and Bradford S. Jones. 2004. \emph{Event History Modeling: A Guide for Social Scientists}. Cambridge University Press.

Butts, Carter T. 2008. "A Relational Event Framework for Social Action." \emph{Sociological Methodology} 38(1): 155-200.

Duxbury, Scott. 2020. \emph{Longitudinal Network Models}. Sage University Press. Quantitative Applications in
the Social Sciences: 192.

Greene, William H. 2003. \emph{Econometric Analysis}. Fifth Edition. Prentice Hall Press.
}
\seealso{
\code{\link{predict.dream_rem}}, \code{\link{vcov.dream_rem}}, \code{\link{logLik.dream_rem}},
\code{\link{AIC}}, \code{\link{gof_rem}}, \code{\link{residuals.dream_rem}},\code{\link{plot.dream_rem}},\code{\link{coef.dream_rem}}.
}
\author{
Kevin A. Carson \href{mailto:kacarson@arizona.edu}{kacarson@arizona.edu} and Diego F. Leal \href{mailto:dflc@arizona.edu}{dflc@arizona.edu}
}
