% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{plot.dream_rem}
\alias{plot.dream_rem}
\title{Plot method for \code{dream_rem} Relational Event Model Fits}
\usage{
\method{plot}{dream_rem}(x, type = c("std.deviance", "schoenfeld"), ...)
}
\arguments{
\item{x}{An object of class "dream_rem".}

\item{type}{If "std.deviance", the returned plot is the standardized deviance
residual values (y) by realized event times (x). If "schoenfeld", then the
returned plots are for each covariates, (2) the Schoenfeld residual values (y) by realized event times (x).}

\item{...}{Additional arguments for other methods.}
}
\description{
Plot the residuals of a \code{dream_rem} relational event model fit. Currently,
the \code{plot} function returns plots of either: (1) standardized deviance
residual values (y) by realized event times (x) or, for each
covariate, (2) the Schoenfeld residual values (y) by realized event times (x).
}
\details{
Generate plots for \code{dream_rem} relational event model fits to plot
model diagnostics.
}
\examples{
#Creating a psuedo one-mode relational event sequence with ordinal timing
relational.seq <- simulate_rem_seq(n_actors = 8,
                                  n_events = 50,
                                  inertia = TRUE,
                                  inertia_p = 0.10,
                                  sender_outdegree = TRUE,
                                  sender_outdegree_p = 0.05)

#Creating a post-processing event sequence for the above relational sequence
post.processing <-  create_res(type = "one-mode",
                              ordinal = TRUE,
                               riskset = "constant_sample",
                              time = relational.seq$eventID,
                              sender = as.character(relational.seq$sender),
                              receiver = as.character(relational.seq$target),
                              n_controls = 5)

#Computing the sender-outdegree statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_degree(formation = "sender-outdegree",
                                  data = post.processing,
                                  halflife = 2)

#Computing the inertia/repetition statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_repetition(data = post.processing,
                                        halflife = 2)

#Fitting an ordinal timing relational event model to the above one-mode relational
#event sequence
rem <- estimate_rem(~ sender.outdegree + repetition,
                         data=post.processing)
summary(rem) #summary of the relational event model

#plotting the standardized deviance residuals for the estimated model
plot(rem, type="std.deviance")
}
