% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream_estimate.R
\name{vcov.dream_rem}
\alias{vcov.dream_rem}
\title{Extract variance-covariance matrix from Relational Event Model Fits}
\usage{
\method{vcov}{dream_rem}(object, ...)
}
\arguments{
\item{object}{An object of class "dream_rem".}

\item{...}{Additional arguments for other methods.}
}
\description{
This function extracts the variance-covariance matrix from estimated
relational event model fits.
}
\examples{
#Creating a psuedo one-mode relational event sequence with ordinal timing
relational.seq <- simulate_rem_seq(n_actors = 8,
                                  n_events = 50,
                                  inertia = TRUE,
                                  inertia_p = 0.10,
                                  sender_outdegree = TRUE,
                                  sender_outdegree_p = 0.05)

#Creating a post-processing event sequence for the above relational sequence
post.processing <-  create_res(type = "one-mode",
                              riskset = "constant_sample",
                              ordinal = TRUE,
                              time = relational.seq$eventID,
                              sender = as.character(relational.seq$sender),
                              receiver = as.character(relational.seq$target),
                              n_controls = 5)

#Computing the sender-outdegree statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_degree(formation = "sender-outdegree",
                                  data = post.processing,
                                  halflife = 2)

#Computing the inertia/repetition statistic for the above post-processing
#one-mode relational event sequence
post.processing <- dreamstats_repetition(data = post.processing,
                                        halflife = 2)

#Fitting an ordinal timing relational event model to the above one-mode relational
#event sequence
rem <- estimate_rem(~ sender.outdegree + repetition,
                         data=post.processing)
vcov(rem)

}
